### 2024-03-19 #632 @ldecicco-USGS
Thanks for the submission @giovsaraceno ! I'm getting some advice from the other editors about your question. One thing that would be really helpful - could you push up your documentation to a GitHub page?

From the `usethis` package, there's a function that helps setting it up:
https://usethis.r-lib.org/reference/use_github_pages.html
