### 2024-03-20 #632 @mpadge
Hi @giovsaraceno, Mark here from the rOpenSci stats team to answer your question. We've done our best to clarify the role of [*Probability Distributions* Standards](https://stats-devguide.ropensci.org/standards.html#standards-distributions):

> Unlike most other categories of standards, packages which fit in this category will also generally be expected to fit into at least one other category of statistical software. Reflecting that expectation, standards for probability distributions will be expected to only pertain to some (potentially small) portion of code in any package.

So packages should generally fit within some main category, with *Probability Distributions* being an additional category. In your case, *Dimensionality Reduction* seems like the appropriate main category, but it seems like your package would also fit within *Probability Distributions.* Given that, the next step would be for you to estimate what proportion of those standards you think might apply to your package? Our general rule-of-thumb is that at least 50% should apply, but for *Probability Distributions* as an additional category, that figure may be lower.

We are particularly keen to document compliance with this category, because it is where our standards have a large overlap with many core routines of the R language itself. As always, we encourage feedback on our standards, so please also feel very welcome to open issues in the [*Stats Software* repository](https://github.com/ropensci/statistical-software-review-book), or add comments or questions in [the discussion pages](https://github.com/ropensci/statistical-software-review-book/discussions). Thanks for you submission!
