+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: ~
+url: https://giovsaraceno.github.io/QuadratiK-package/
 template:
   bootstrap: 5
 
