+++ DESCRIPTION
@@ -20,7 +20,7 @@ Description: It includes test for multivariate normality, test for
     <doi:10.5705/ss.202022.0347>, and Golzy, M., Markatou, M. (2020)
     <doi:10.1080/10618600.2020.1740713>.
 License: GPL (>= 3)
-URL: https://cran.r-project.org/web/packages/QuadratiK/index.html, https://github.com/giovsaraceno/QuadratiK-package
+URL: https://cran.r-project.org/web/packages/QuadratiK/index.html, https://github.com/giovsaraceno/QuadratiK-package, https://giovsaraceno.github.io/QuadratiK-package/
 BugReports: https://github.com/giovsaraceno/QuadratiK-package/issues
 Depends: 
     R (>= 3.5.0)
