+++ man/QuadratiK-package.Rd
@@ -35,9 +35,8 @@ Science Foundation and Department of Biostatistics, University at Buffalo.
 }
 \references{
 Saraceno Giovanni, Markatou Marianthi, Mukhopadhyay Raktim, Golzy Mojgan
-(2023). Goodness-of-fit and clustering of spherical data: The QuadratiK
-package in R and Python. Technical Report, Department of Biostatistics,
-University at Buffalo.
+(2024). Goodness-of-Fit and Clustering of Spherical Data: the QuadratiK
+package in R and Python. arXiv preprint arXiv:2402.02290.
 
 Ding Yuxin, Markatou Marianthi, Saraceno Giovanni (2023). “Poisson
 Kernel-Based Tests for Uniformity on the d-Dimensional Sphere.”
