+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://giovsaraceno.github.io/QuadratiK-package/
+url: ~
 template:
   bootstrap: 5
 
