### 2024-03-22 3f6af55
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@6b82650e3b77b36261eab65ace0ca01005dd87eb 🚀
