+++ pkgdown.yml
@@ -1,0 +1,10 @@
+pandoc: 3.1.11
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  TwoSample_test: TwoSample_test.html
+  kSample_test: kSample_test.html
+  uniformity: uniformity.html
+  wireless_clustering: wireless_clustering.html
+last_built: 2024-03-22T17:03Z
+
