### 2024-03-22 #632 @giovsaraceno
> Hi @giovsaraceno, Mark here from the rOpenSci stats team to answer your question. We've done our best to clarify the role of [_Probability Distributions_ Standards](https://stats-devguide.ropensci.org/standards.html#standards-distributions):
>
> > Unlike most other categories of standards, packages which fit in this category will also generally be expected to fit into at least one other category of statistical software. Reflecting that expectation, standards for probability distributions will be expected to only pertain to some (potentially small) portion of code in any package.
>
> So packages should generally fit within some main category, with _Probability Distributions_ being an additional category. In your case, _Dimensionality Reduction_ seems like the appropriate main category, but it seems like your package would also fit within _Probability Distributions._ Given that, the next step would be for you to estimate what proportion of those standards you think might apply to your package? Our general rule-of-thumb is that at least 50% should apply, but for _Probability Distributions_ as an additional category, that figure may be lower.
>
> We are particularly keen to document compliance with this category, because it is where our standards have a large overlap with many core routines of the R language itself. As always, we encourage feedback on our standards, so please also feel very welcome to open issues in the [_Stats Software_ repository](https://github.com/ropensci/statistical-software-review-book), or add comments or questions in [the discussion pages](https://github.com/ropensci/statistical-software-review-book/discussions). Thanks for you submission!

Hi Mark,

Thank you for the additional clarification regarding the standards for Probability Distributions and their integration with other statistical software categories. Following your guidance, we have conducted a thorough review of the standards applicable to the Probability Distributions category in relation to our package.

Based on our assessment, we found that the current version of our package satisfies 14% of the standards directly. Furthermore, we identified that an additional 36% of the standards could potentially apply to our package, but this would require us to make some enhancements, including the addition of checks and test codes. We feel the remaining  50% of the standards are not applicable to our package.

We are committed to improve our package and aim to fulfill the applicable standards. To this end, we plan to work on a separate branch dedicated to implementing these enhancements, with the goal of meeting the 50% of the standards for the Probability Distributions category. Before proceeding, we would greatly appreciate your opinion on this plan.

Thank you for your time and support. Giovanni
