### 2024-03-23 1fe0b65
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@1e5ef9f249c61414c55f28a2c3485750149a0374 🚀
