### 2024-03-23 32842ac
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@74eea6cad55a70209b4b0d4f9cb85f3550ea63ff 🚀
