+++ R/pkbd_functions.R
@@ -94,6 +94,14 @@ dpkb <- function(x, mu, rho, logdens = FALSE) {
 #'    \item \code{'rejpsaw'} using projected Saw distributions 
 #'                         (Algorithm in Table 2 of Sablica et al. 2023).
 #' }
+#' @param max.iter the maximum number of iterations (for 'rejacg' method).
+#' @param tol.eps the desired accuracy of convergence tolerance 
+#'               (for 'rejacg' method).
+#' 
+#' @details
+#' If the chosen method is 'rejacg', the function \code{uniroot}, from the 
+#' \code{stat} package, is used to estimate the beta parameter. In this case, 
+#' the complete results are provided as output. 
 #' 
 #' @references
 #' Golzy, M., Markatou, M. (2020) Poisson Kernel-Based Clustering on the Sphere:
@@ -107,12 +115,22 @@ dpkb <- function(x, mu, rho, logdens = FALSE) {
 #' @import movMF
 #' @import Tinflex
 #'
-#' @srrstats {G1.0} Reference section reports the related literature
-#' @srrstats {G1.4} roxigen2 is used
-#' @srrstats {G2.0,G2.0a} check input mu 
+#' @srrstats {G1.0} Reference section reports the related literature.
+#' @srrstats {G1.4} roxigen2 is used.
+#' @srrstats {G2.0,G2.0a} check input mu. 
+#' @srrstats {PD1.0} references on choice and usage of the probability 
+#'                   distributions are provided.
+#' @srrstats {PD3.2} all parameters of uniroot function are specified
+#'                   (for method 'rejacg').
+#' @srrstats {PD3.3} the complete results of uniroot function are provided as 
+#'                   output (for method 'rejacg').
+#' @srrstats {PD3.1} the three algorithms for random generation are implemented
+#'                   as separate functions.
 #' 
 #' @export
-rpkb <- function(n, mu, rho, method = 'rejvmf') {
+rpkb <- function(n, mu, rho, method = 'rejvmf', 
+                 tol.eps = .Machine$double.eps^0.25, 
+                 max.iter = 1000) {
    
    if (rho >= 1 | rho < 0) {
       stop('Input argument rho must be within [0,1)')
@@ -134,107 +152,187 @@ rpkb <- function(n, mu, rho, method = 'rejvmf') {
    
    mu <- mu / sqrt(sum(mu^2))
    p <- length(mu)
+   
    if (method == 'rejvmf') {
-      kappa<-p*rho/(1+rho^2)
-      M1 <- (1 + rho) / ((1 - rho)^(p-1)*exp(kappa))
-      retvals <- matrix(nrow = n, ncol = p)
-      numAccepted <- 0
-      numTries <- 0
-      while (numAccepted < n) {
-         numTries <- numTries + 1
-         theta<-kappa*mu
-         yx<- rmovMF(1,theta,1)
-         v<- yx%*%mu
-         f1<- (1-rho**2)/ (( 1 + rho**2 - 2*rho*v)^(p/2))
-         u <- runif(1)
-         if (u <= f1/(M1 * exp(kappa*v))) {
-            numAccepted <- numAccepted + 1
-            retvals[numAccepted,] <- yx
-         }
-      }
+      
+      res <- rejvmf(n, rho, mu, p)
       
    } else if (method == 'rejacg') {
       
-      lambda <- (2*rho)/(1+rho^2)
[…]
