### 2024-03-27 #632 @mpadge
Hi Giovanni, your `srrstats` tags for probability distribution standards definitely look good enough to proceed. That said, one aspect which could be improved, and which I would request if I were reviewing the package, is the compliance statements in the tests. In both `test-dpkb.R` and `test-rkpb.R` you claim compliance in single statements at the start, yet I can't really see where or how a few of these are really complied with. In particular, there do not appear to be explicit tests for output *values*, as these are commonly tested using `test_equal` with an explicit `tolerance` parameter, which you don't have. It is also not clear to me where and how you compare results of different distributions, because you have no annotations in the tests about what the return values of the functions are.

Those are very minor points which you may ignore for the moment if you'd like to get the review process started, or you could quickly address them straight away if you prefer. Either way, feel free to ask the bot to `check srr` when you think you're ready to proceed. Thanks!
