+++ .Rproj.user/EE77B124/compilation-database/mingw64/c-definitions.h
@@ -1,1568 +1,0 @@
-#ifndef __DBL_MIN_EXP__
-# define __DBL_MIN_EXP__ (-1021)
-#endif
-
-#ifndef __UINT_LEAST16_MAX__
-# define __UINT_LEAST16_MAX__ 0xffff
-#endif
-
-#ifndef __FLT16_HAS_QUIET_NAN__
-# define __FLT16_HAS_QUIET_NAN__ 1
-#endif
-
-#ifndef __ATOMIC_ACQUIRE
-# define __ATOMIC_ACQUIRE 2
-#endif
-
-#ifndef __FLT128_MAX_10_EXP__
-# define __FLT128_MAX_10_EXP__ 4932
-#endif
-
-#ifndef __FLT_MIN__
-# define __FLT_MIN__ 1.17549435082228750796873653722224568e-38F
-#endif
-
-#ifndef __GCC_IEC_559_COMPLEX
-# define __GCC_IEC_559_COMPLEX 2
-#endif
-
-#ifndef __UINT_LEAST8_TYPE__
-# define __UINT_LEAST8_TYPE__ unsigned char
-#endif
-
-#ifndef __SIZEOF_FLOAT80__
-# define __SIZEOF_FLOAT80__ 16
-#endif
-
-#ifndef __INTMAX_C
-# define __INTMAX_C(c) c ## LL
-#endif
-
-#ifndef __CHAR_BIT__
-# define __CHAR_BIT__ 8
-#endif
-
-#ifndef __MINGW32__
-# define __MINGW32__ 1
-#endif
-
-#ifndef __UINT8_MAX__
-# define __UINT8_MAX__ 0xff
-#endif
-
-#ifndef __SCHAR_WIDTH__
-# define __SCHAR_WIDTH__ 8
-#endif
-
-#ifndef _WIN64
-# define _WIN64 1
-#endif
-
-#ifndef __WINT_MAX__
-# define __WINT_MAX__ 0xffff
-#endif
-
-#ifndef __FLT32_MIN_EXP__
-# define __FLT32_MIN_EXP__ (-125)
-#endif
-
-#ifndef __ORDER_LITTLE_ENDIAN__
-# define __ORDER_LITTLE_ENDIAN__ 1234
-#endif
-
-#ifndef __SIZE_MAX__
-# define __SIZE_MAX__ 0xffffffffffffffffULL
-#endif
-
-#ifndef __WCHAR_MAX__
-# define __WCHAR_MAX__ 0xffff
-#endif
-
-#ifndef __GCC_HAVE_SYNC_COMPARE_AND_SWAP_2
-# define __GCC_HAVE_SYNC_COMPARE_AND_SWAP_2 1
-#endif
-
-#ifndef __GCC_HAVE_SYNC_COMPARE_AND_SWAP_4
-# define __GCC_HAVE_SYNC_COMPARE_AND_SWAP_4 1
-#endif
-
-#ifndef __GCC_HAVE_SYNC_COMPARE_AND_SWAP_8
-# define __GCC_HAVE_SYNC_COMPARE_AND_SWAP_8 1
-#endif
-
-#ifndef __GCC_ATOMIC_CHAR_LOCK_FREE
-# define __GCC_ATOMIC_CHAR_LOCK_FREE 2
-#endif
-
-#ifndef __GCC_IEC_559
-# define __GCC_IEC_559 2
-#endif
-
-#ifndef __FLT32X_DECIMAL_DIG__
-# define __FLT32X_DECIMAL_DIG__ 17
-#endif
-
-#ifndef __FLT_EVAL_METHOD__
-# define __FLT_EVAL_METHOD__ 0
-#endif
-
[…]
