### 2024-03-27 9727306
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@3c1a737b133718e0477a7e29dab08cbec214c7c7 🚀
