### 2024-03-27 c51bbd8
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@f7f3432e3c692c352b2c9166139b9b75be174f94 🚀
