+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   kSample_test: kSample_test.html
   uniformity: uniformity.html
   wireless_clustering: wireless_clustering.html
-last_built: 2024-03-27T14:06Z
+last_built: 2024-03-27T14:09Z
 
