### 2024-03-28 e57a4ca
new vignette
Vignette about the random sampling from PKBD
