### 2024-03-28 0d1cd2b
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@e57a4ca264fc397789ea8ec123cf4e140e70014e 🚀
