+++ pkgdown.yml
@@ -3,8 +3,9 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   TwoSample_test: TwoSample_test.html
+  generate_rpkb: generate_rpkb.html
   kSample_test: kSample_test.html
   uniformity: uniformity.html
   wireless_clustering: wireless_clustering.html
-last_built: 2024-03-27T14:09Z
+last_built: 2024-03-28T18:25Z
 
