### 2024-03-28 6a2684a
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@e3ca11e0c0d8cdf244bef491e735b0ab67df543e 🚀
