### 2024-03-28 4da30b6
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@56cb0092caf291950de7fee0b8997726f2241bc8 🚀
