### 2024-03-28 a969555
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@1b536675736f155f44ede1d779e17aabb026e112 🚀
