+++ pkgdown.yml
@@ -7,5 +7,5 @@ articles:
   kSample_test: kSample_test.html
   uniformity: uniformity.html
   wireless_clustering: wireless_clustering.html
-last_built: 2024-03-28T19:36Z
+last_built: 2024-03-28T19:40Z
 
