### 2024-03-28 3d88e57
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@190208dcb21d2107a788e2bac7954634350f2874 🚀
