### 2024-04-02 1defa38
kernel computation
Adjustments for the steps of the kernel tests computation.
