### 2024-04-04 dd669e2
V-statistics normality
