+++ R/RcppExports.R
@@ -109,8 +109,7 @@ stat2sample <- function(x_mat, y_mat, h, mu_hat, Sigma_hat, centeringType = "Non
 #' @param h The bandwidth parameter for the Gaussian kernel function.
 #' @param mu_hat Mean vector for the reference distribution (if available)
 #' @param Sigma_hat Covariance matrix of the reference distribution (if available)
-#' @param centeringType String indicating the method used for centering the normal kernel ('Param' or 'Nonparam').
-#'
+#' 
 #' @return A scalar value representing the test statistic.
 #'
 #' @useDynLib QuadratiK
@@ -118,8 +117,8 @@ stat2sample <- function(x_mat, y_mat, h, mu_hat, Sigma_hat, centeringType = "Non
 #' @keywords internal
 #' 
 #' @noRd
-kbNormTest <- function(x_mat, h, mu_hat, Sigma_hat, centeringType = "Param") {
-    .Call('_QuadratiK_kbNormTest', PACKAGE = 'QuadratiK', x_mat, h, mu_hat, Sigma_hat, centeringType)
+kbNormTest <- function(x_mat, h, mu_hat, Sigma_hat) {
+    .Call('_QuadratiK_kbNormTest', PACKAGE = 'QuadratiK', x_mat, h, mu_hat, Sigma_hat)
 }
 
 #' Poisson kernel-based test for Uniformity on the Sphere
