+++ R/h_selection.R
@@ -210,8 +210,8 @@ select_h <- function(x, y=NULL, alternative=NULL, method="subsampling", b=0.8,
       #x <- as.matrix(xnew[sample(n, replace = FALSE),])
       #y <- as.matrix(ynew[sample(m, replace = FALSE),])
       
-      STATISTIC <- stat2sample(xnew, ynew, h, matrix(0,nrow=1),
-                               diag(1),"Nonparam")
+      STATISTIC <- stat2sample(xnew, ynew, h, rep(0,d),
+                               diag(d),"Nonparam")
       CV <- compute_CV(B, Quantile, pooled, n, m, h, method, b)
       
       return(STATISTIC < CV)
@@ -269,9 +269,8 @@ select_h <- function(x, y=NULL, alternative=NULL, method="subsampling", b=0.8,
       
       xnew <- sn::rmsn(n, xi = mean_tilde, Omega = S_tilde, alpha = skew_tilde)
       
-      STATISTIC <- kbNormTest(xnew, h, matrix(mean_dat,nrow = 1), S_dat, 
-                              centeringType = "Param")
-      CV <- normal_CV(d,n,h,matrix(mean_dat,nrow = 1),S_dat,B,Quantile)
+      STATISTIC <- kbNormTest(xnew, h, mean_dat, S_dat)
+      CV <- normal_CV(d,n,h,mean_dat,S_dat,B,Quantile)
       
       return(c(STATISTIC < CV))
    }

[…]
