### 2024-04-10 #632 @noamross
Sorry we didn't reply faster, @giovsaraceno. In, say, a single-variable distribution tests might include:
 - A correctness that the density function with given parameters has means, modes, or variances as theoretically expected.
 - A parameter recovery that the mean of a sufficiently large number of randomly generated values is within a window of expectations.
In your case my understanding is that you are generating multivariate outputs.  Ultimately we aim to see tests that those outputs are as expected, so for both density and random values.  I think the thing to do is test that summary properties of those outputs, deterministic for density and within bounds for random, match those expected based on the input parameters
