### 2024-04-15 fd6e7b8
test numeric values
numeric tests for dpkb and rpkb
