### 2024-04-15 8bddb1f
Update kernel_function.cpp
output two k-sample statistics
- two-sample style
- trace
