+++ R/pk.test.R
@@ -109,6 +109,7 @@ setMethod("pk.test", signature(x = "ANY"),
              var_Un <- (2/(n*(n-1)))*((1+rho^2)/((1-rho^2)^(d-1)) -1)
              
              CV_Un <- poisson_CV(d=d, size=n, rho=rho, B=B, Quantile=Quantile )
+             CV_Un <- CV_Un/sqrt(var_Un)
              
              res <- new("pk.test", Un = pk[1]/sqrt(var_Un), CV_Un = CV_Un, 
                         Vn = pk[2], CV_Vn = CV_Vn, method = METHOD, x = x, B= B,
