### 2024-04-26 6ee08b0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@e54e445114d2d4ce2b7c7ebe262404376baaafe1 🚀
