+++ pkgdown.yml
@@ -1,11 +1,11 @@
 pandoc: 3.1.11
-pkgdown: 2.0.7
+pkgdown: 2.0.9
 pkgdown_sha: ~
 articles:
-  TwoSample_test: TwoSample_test.html
   generate_rpkb: generate_rpkb.html
   kSample_test: kSample_test.html
+  TwoSample_test: TwoSample_test.html
   uniformity: uniformity.html
   wireless_clustering: wireless_clustering.html
-last_built: 2024-03-28T20:15Z
+last_built: 2024-04-26T16:42Z
 
