### 2024-04-26 1b010d0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@299f014c820aceb937962d78e3ba7798ab12fd19 🚀
