### 2024-04-26 e5604e2
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@21541a401b227b2808949db0228f5c908cfba6b1 🚀
