+++ pkgdown.yml
@@ -7,5 +7,5 @@ articles:
   TwoSample_test: TwoSample_test.html
   uniformity: uniformity.html
   wireless_clustering: wireless_clustering.html
-last_built: 2024-04-26T16:45Z
+last_built: 2024-04-26T16:48Z
 
