### 2024-05-01 5bcc42d
corrections of the provided statistics
- compute Un and Vn for the normality test with corresponding asymptotic
- compute also trace statistic for two sample
- compute exact variance for two sample
- compute sample variance for trace and k-sample statistics
