### 2024-05-01 33882de
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@4f9511b2c651db3f9a03a2593f974231bc24cccc 🚀
