### 2024-05-01 4dd3718
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@1fffc2e4785e6d57a20fb911b49aab126c756204 🚀
