+++ tests/testthat/test-select_h.R
@@ -53,14 +53,14 @@ test_that("Select h", {
    expect_equal(class(result$power), "data.frame")
    
    # two-sample
-   result <- select_h(x = matrix(rnorm(20),ncol=2), 
-                      y = matrix(rnorm(20),ncol=2), alternative="skewness")
+   result <- select_h(x = matrix(rnorm(20),ncol=2),
+   y = matrix(rnorm(20),ncol=2), alternative="skewness")
    expect_equal(class(result$h_sel), "numeric")
    expect_equal(class(result$power), "data.frame")
    
    # k-sample
-   result <- select_h(x = matrix(rnorm(30),ncol=2), y = rep(c(1,2,3),each=5), 
-                      alternative="scale")
+   result <- select_h(x = matrix(rnorm(30),ncol=2), y = rep(c(1,2,3),each=5),
+   alternative="scale")
    expect_equal(class(result$h_sel), "numeric")
    expect_equal(class(result$power), "data.frame")
 })
