### 2024-05-02 71febc5
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@777189fa8ba3bb1f691caed2dda54a85b23f8524 🚀
