+++ DESCRIPTION
@@ -23,7 +23,7 @@ License: GPL (>= 3)
 URL: https://cran.r-project.org/web/packages/QuadratiK/index.html, https://github.com/giovsaraceno/QuadratiK-package, https://giovsaraceno.github.io/QuadratiK-package/
 BugReports: https://github.com/giovsaraceno/QuadratiK-package/issues
 Depends: 
-    R (>= 4.4.0)
+    R (>= 3.6.0)
 Imports: 
     cluster,
     clusterRepro,
@@ -32,7 +32,7 @@ Imports:
     ggplot2,
     ggpp,
     ggpubr,
-    MASS,
+    MASS (>= 7.3-54),
     mclust,
     methods,
     moments,
