### 2024-05-02 7fa803c
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@3fd69d87e36c69800f269b5e123a74e5867950e2 🚀
