### 2024-05-03 ebd312e
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@dcf867646910d9400d01a42d34827dd0e0232a9a 🚀
