### 2024-05-03 170d3c8
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@0920e1acdc424fbbcb3bac700de382a4563e5e59 🚀
