+++ .github/workflows/pkgcheck.yaml
@@ -16,7 +16,7 @@ on:
 
 jobs:
   pkgcheck:
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
     permissions:
       issues: write
     env:
