### 2024-05-03 7890238
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@69b67286126d9815e11237288870d58febd46a9d 🚀
