+++ .github/workflows/pkgcheck.yaml
@@ -22,8 +22,4 @@ jobs:
     env:
       RGL_USE_NULL: TRUE
     steps:
-      - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: "4.4.0"
       - uses: ropensci-review-tools/pkgcheck-action@main
