### 2024-05-03 6fc4b2e
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@0313fec75c17b2b8957962a54c3a20afac0f5894 🚀
