+++ .github/workflows/pkgcheck.yaml
@@ -16,7 +16,7 @@ on:
 
 jobs:
   pkgcheck:
-    runs-on: windows-latest
+    runs-on: macos-latest
     permissions:
       issues: write
     env:
