+++ .github/workflows/pkgcheck.yaml
@@ -16,10 +16,12 @@ on:
 
 jobs:
   pkgcheck:
-    runs-on: macos-latest
+    runs-on: ubuntu-latest
     permissions:
       issues: write
     env:
       RGL_USE_NULL: TRUE
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
+        with:
+          r-version: "4.4.0"
