### 2024-05-03 eab8165
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@679d5d33b81ac2ae250d9c5deff6027fc232f1de 🚀
