### 2024-05-03 4d8f2b1
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@ebcce47b09453230313638de61f09bd733952e81 🚀
