### 2024-05-03 a36a9f0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@64d0677e42b76eb79573e7d5d241b9c747c534f0 🚀
