### 2024-05-03 d9ce570
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@4987e56b200bb5d2e5077455d77f19fe62d0d943 🚀
