### 2024-05-03 40512a0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@ba76904de88eca849d3c90a0ed15e6ff3ada9edc 🚀
