### 2024-05-03 8c6d8c9
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@d52ecfe67810b56e637114602738029af85ac02e 🚀
