### 2024-05-03 7300f73
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@ec4d09bccfca4fbde4866a59e1ae0bcd9b70a8c7 🚀
