+++ NAMESPACE
@@ -34,7 +34,6 @@ import(mvtnorm)
 import(rgl)
 import(rlecuyer)
 import(stats)
-importFrom(MASS,ginv)
 importFrom(Rcpp,sourceCpp)
 importFrom(cluster,silhouette)
 importFrom(grDevices,colorRampPalette)
