### 2024-05-03 e839286
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@efb54045009b63f27ee66c8b941761dc81d70b6d 🚀
