+++ .github/workflows/pkgcheck.yaml
@@ -18,5 +18,9 @@ jobs:
   pkgcheck:
     runs-on: ubuntu-latest
     steps:
-      - uses: ropensci-review-tools/pkgcheck-action@main
-
+    - name: Set up R
+      uses: r-lib/actions/setup-r@v2
+      with:
+        r-version: '4.4.0'
+    - uses: ropensci-review-tools/pkgcheck-action@main
+    
