### 2024-05-03 9568bd7
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@e1b47554345a4c6f4655ae3fa6e6c2e8742a5672 🚀
