+++ .github/workflows/pkgcheck.yaml
@@ -17,6 +17,8 @@ on:
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
+    permissions:
+      issues: write
     steps:
     - name: Set up R
       uses: r-lib/actions/setup-r@v2
