### 2024-05-03 3c33e88
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@7abee577fcb399f9246ccf365f925d889e81be3f 🚀
