+++ .github/workflows/pkgcheck.yaml
@@ -20,6 +20,12 @@ jobs:
     permissions:
       issues: write
     steps:
+    - name: Check R version
+      run: R --version
+    - name: Install pak and update packages
+      run: |
+        install.packages("pak", repos = "https://cran.r-project.org")
+        pak::pkg_install(ask = FALSE, verbose = TRUE)
     - name: Set up R
       uses: r-lib/actions/setup-r@v2
       with:
