### 2024-05-03 10f2524
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@f9d8df3aca2bb0f4f9a89497981b48255693456f 🚀
