### 2024-05-03 b380870
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@f64ab2e1c52f7c1961f1791a4c8bda87ee010a4f 🚀
