### 2024-05-03 7bf791a
Update pkgcheck.yaml
