+++ .github/workflows/pkgcheck.yaml
@@ -26,5 +26,7 @@ jobs:
       uses: r-lib/actions/setup-r@v2
       with:
         r-version: '4.4.0'
+    - name: Check R version
+      run: R --version
     - uses: ropensci-review-tools/pkgcheck-action@main
     
