### 2024-05-03 35f3594
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@7bf791a28f8b09eed9528f8fa145d6bdbb622cb4 🚀
