+++ .github/workflows/pkgcheck.yaml
@@ -23,10 +23,11 @@ jobs:
     - name: Check R version
       run: R --version
     - name: Set up R
-      uses: r-lib/actions/setup-r@v2
+      uses: |
+        r-lib/actions/setup-r@v2
+        ropensci-review-tools/pkgcheck-action@main
       with:
         r-version: '4.4.0'
     - name: Check R version
       run: R --version
-    - uses: ropensci-review-tools/pkgcheck-action@main
-    
+     
