+++ .github/workflows/pkgcheck.yaml
@@ -23,11 +23,12 @@ jobs:
     - name: Check R version
       run: R --version
     - name: Set up R
-      uses: |
-        r-lib/actions/setup-r@v2
-        ropensci-review-tools/pkgcheck-action@main
+      uses: r-lib/actions/setup-r@v2
       with:
         r-version: '4.4.0'
     - name: Check R version
       run: R --version
+    - uses: ropensci-review-tools/pkgcheck-action@main
+      with:
+        r-version: '4.4.0'
      
