### 2024-05-03 8427a00
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@d0d674bbcb27e670960d8dc5eac4b1c30d037833 🚀
