### 2024-05-03 8926564
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@abdc3fb15e4c3549a76227168cdce749da7f4683 🚀
