### 2024-05-03 5570a8e
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@31f50ec6a307c9bab4eba9b7eba9d3cf6bffe579 🚀
