### 2024-05-04 89c6974
addition of general formulas
