+++ src/kernel_function.cpp
@@ -150,7 +150,7 @@ Eigen::MatrixXd ParamCentering(const Eigen::MatrixXd& kmat_zz, const Eigen::Matr
 //' @srrstats {G1.4a} roxigen2 is used
 //' @keywords internal
 // [[Rcpp::export]]
-double var_two(const Eigen::MatrixXd& Kcen, const Eigen::VectorXd& n_samples) 
+Eigen::VectorXd var_two(const Eigen::MatrixXd& Kcen, const Eigen::VectorXd& n_samples) 
    {
    
    int n = n_samples(0);
@@ -168,19 +168,26 @@ double var_two(const Eigen::MatrixXd& Kcen, const Eigen::VectorXd& n_samples)
    double m_factor = 1.0 / (m * (m - 1));
    double cross_factor = 1.0 / (n * m);
    
-   // Variance estimate calculation
-   double est_var = 2 * n_factor * n_factor * K_xx.array().square().sum() +
-      8 * cross_factor * cross_factor * K_xy.array().square().sum() +
-      2 * m_factor * m_factor * K_yy.array().square().sum();
+   // Variance estimate calculation - Dn
+   double est_varD = 2 * n_factor * n_factor * pow(K_xx.array(),2).sum() +
+      8 * cross_factor * cross_factor * pow(K_xy.array(),2).sum() +
+      2 * m_factor * m_factor * pow(K_yy.array(),2).sum();
    
+   // Variance estimate calculation - trace
+   double est_var_Tr = 2 * n_factor * n_factor * pow(K_xx.array(),2).sum() +
+      2 * m_factor * m_factor * pow(K_yy.array(),2).sum();
    
    double delta1 = (K_xx * K_xy.transpose()).sum();
    double delta2 = (K_xy * K_yy.transpose()).sum();
    
-   est_var -= 8 * n_factor * cross_factor * delta1;
-   est_var -= 8 * m_factor * cross_factor * delta2;
+   est_varD -= 8 * n_factor * cross_factor * delta1;
+   est_varD -= 8 * m_factor * cross_factor * delta2;
    
-   return est_var;
+   Eigen::VectorXd results(2);
+   results(0) = est_varD;
+   results(1) = est_var_Tr;
+   
+   return results;
 }
 //' Compute kernel-based quadratic distance two-sample test with Normal kernel
 //'
@@ -238,7 +245,7 @@ Eigen::VectorXd stat2sample(Eigen::MatrixXd& x_mat, Eigen::MatrixXd& y_mat, doub
    n_sample(0) = n_x;
    n_sample(1) = n_y;
       
-   double var_nonparam = var_two(k_center, n_sample);
+   Eigen::VectorXd var_nonparam = var_two(k_center, n_sample);
       
    k_center.diagonal().setZero();
    
@@ -249,10 +256,11 @@ Eigen::VectorXd stat2sample(Eigen::MatrixXd& x_mat, Eigen::MatrixXd& y_mat, doub
    double Test_trace = (k_center.block(0, 0, n_x, n_x).sum() / (n_x * (n_x - 1)))  +
       (k_center.block(n_x, n_x, n_y, n_y).sum() / (n_y * (n_y - 1)));
    
-   Eigen::VectorXd results(3);
+   Eigen::VectorXd results(4);
    results(0) = n_z * Test_NonPar;
    results(1) = n_z * Test_trace;
-   results(2) = var_nonparam;
+   results(2) = var_nonparam(0);
+   results(3) = var_nonparam(1);
    return results;
 }
 //' Compute kernel-based quadratic distance test for Normality
@@ -340,6 +348,69 @@ Eigen::VectorXd statPoissonUnif(const Eigen::MatrixXd& x_mat, double rho)
    
[…]
