### 2024-05-04 359cd8c
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@89c69749bfea3a9530749abbc5bd315d5da60ee3 🚀
