### 2024-05-05 838decb
updates
- change estract_stats in stats_clusters
- set R 3.6.0 as required
- fix tags in documentations
