+++ NAMESPACE
@@ -12,7 +12,6 @@ export(select_h)
 exportClasses(kb.test)
 exportClasses(pk.test)
 exportClasses(pkbc)
-exportMethods(extract_stats)
 exportMethods(kb.test)
 exportMethods(pk.test)
 exportMethods(pkbc)

[…]
