+++ src/RcppExports.cpp
@@ -63,34 +63,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// var_two
-Eigen::VectorXd var_two(const Eigen::MatrixXd& Kcen, const Eigen::VectorXd& n_samples);
-RcppExport SEXP _QuadratiK_var_two(SEXP KcenSEXP, SEXP n_samplesSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Kcen(KcenSEXP);
-    Rcpp::traits::input_parameter< const Eigen::VectorXd& >::type n_samples(n_samplesSEXP);
-    rcpp_result_gen = Rcpp::wrap(var_two(Kcen, n_samples));
-    return rcpp_result_gen;
-END_RCPP
-}
-// stat2sample
-Eigen::VectorXd stat2sample(Eigen::MatrixXd& x_mat, Eigen::MatrixXd& y_mat, double h, const Eigen::VectorXd& mu_hat, const Eigen::MatrixXd& Sigma_hat, const std::string& centeringType);
-RcppExport SEXP _QuadratiK_stat2sample(SEXP x_matSEXP, SEXP y_matSEXP, SEXP hSEXP, SEXP mu_hatSEXP, SEXP Sigma_hatSEXP, SEXP centeringTypeSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Eigen::MatrixXd& >::type x_mat(x_matSEXP);
-    Rcpp::traits::input_parameter< Eigen::MatrixXd& >::type y_mat(y_matSEXP);
-    Rcpp::traits::input_parameter< double >::type h(hSEXP);
-    Rcpp::traits::input_parameter< const Eigen::VectorXd& >::type mu_hat(mu_hatSEXP);
-    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Sigma_hat(Sigma_hatSEXP);
-    Rcpp::traits::input_parameter< const std::string& >::type centeringType(centeringTypeSEXP);
-    rcpp_result_gen = Rcpp::wrap(stat2sample(x_mat, y_mat, h, mu_hat, Sigma_hat, centeringType));
-    return rcpp_result_gen;
-END_RCPP
-}
 // kbNormTest
 Eigen::VectorXd kbNormTest(Eigen::MatrixXd x_mat, double h, const Eigen::VectorXd& mu_hat, const Eigen::MatrixXd Sigma_hat);
 RcppExport SEXP _QuadratiK_kbNormTest(SEXP x_matSEXP, SEXP hSEXP, SEXP mu_hatSEXP, SEXP Sigma_hatSEXP) {
@@ -114,6 +86,34 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type x_mat(x_matSEXP);
     Rcpp::traits::input_parameter< double >::type rho(rhoSEXP);
     rcpp_result_gen = Rcpp::wrap(statPoissonUnif(x_mat, rho));
+    return rcpp_result_gen;
+END_RCPP
+}
+// var_two
+Eigen::VectorXd var_two(const Eigen::MatrixXd& Kcen, const Eigen::VectorXd& nsamples);
+RcppExport SEXP _QuadratiK_var_two(SEXP KcenSEXP, SEXP nsamplesSEXP) {
[…]
