+++ vignettes/wireless_clustering.R
@@ -23,7 +23,7 @@ res_pk <- pkbc(as.matrix(wire_norm),3:6)
 #  round(res_validation$metrics, 8)
 
 ## ----fig.width=6, fig.height=8------------------------------------------------
-summary_clust <- extract_stats(res_pk,4)
+summary_clust <- stats_clusters(res_pk,4)
 
 ## -----------------------------------------------------------------------------
 summary_clust
