### 2024-05-05 5d3a983
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@92d2401cd4235d30ba3144544ce3dd10d311df83 🚀
