### 2024-05-06 9aa8e88
fix variance computation
