### 2024-05-06 8459b46
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@9aa8e884a54c849640f3edfbb707a9eb5423397d 🚀
