### 2024-05-06 0e5bae9
remove ggplot2 and ggpp
these packages are removed since they seem to require the last version of R 4.4.0. The are mantained in the branch "R version 4.4.0".
