+++ man/compute_stats.Rd
@@ -2,23 +2,21 @@
 % Please edit documentation in R/utility.R
 \name{compute_stats}
 \alias{compute_stats}
-\title{Compute and display some descriptive statistics for the two sample tests}
+\title{Compute some descriptive statistics for the two sample tests}
 \usage{
-compute_stats(var1, var2, var_name, eps = 3)
+compute_stats(var1, var2, eps = 3)
 }
 \arguments{
 \item{var1}{vector of observations of a given variable from sample 1}
 
 \item{var2}{vector of observations of a given variable from sample 2}
 
-\item{var_name}{Name of the variable displayed}
-
 \item{eps}{precision of displayed statistics}
 }
 \value{
-Computed statistics with a plot
+Computed statistics
 }
 \description{
-Compute and display some descriptive statistics for the two sample tests
+Compute some descriptive statistics for the two sample tests
 }
 \keyword{internal}

[…]
