### 2024-05-06 429a021
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@0e5bae9e8d74423fb2093db883b42ce9b82bca56 🚀
