### 2024-05-06 de883e7
Update pkgcheck.yaml
