+++ .github/workflows/pkgcheck.yaml
@@ -22,6 +22,14 @@ jobs:
     steps:
     - name: Check R version
       run: R --version
+    - uses: actions/checkout@v2
+    - uses: r-lib/actions/setup-r@v2
+      with:
+        r-version: '4.4.0'  # Specify the latest R version
+    - name: Install dependencies
+      run: |
+        install.packages(c("devtools", "roxygen2"))
+        devtools::install_deps(dependencies = TRUE)
     - uses: ropensci-review-tools/pkgcheck-action@main
       with:
           summary-only: false
