### 2024-05-06 330f68c
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@cfad4f930f2f47b72beee846fee2f9ffa017d9b5 🚀
