### 2024-05-06 39fc428
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@de883e79a7c5ff27ae7efd751245898a0d5d1647 🚀
