+++ .github/workflows/pkgcheck.yaml
@@ -28,8 +28,8 @@ jobs:
         r-version: '4.4.0'  # Specify the latest R version
     - name: Install dependencies
       run: |
-        install.packages(c("devtools", "roxygen2"))
-        devtools::install_deps(dependencies = TRUE)
+        Rscript -e "install.packages(c('devtools', 'roxygen2'))"
+        Rscript -e "devtools::install_deps(dependencies = TRUE)"
     - uses: ropensci-review-tools/pkgcheck-action@main
       with:
           summary-only: false
