### 2024-05-06 bab71b0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@13f4b95e0dbcd70b2fae0a75b84edc0860de1f2f 🚀
