+++ .github/workflows/pkgcheck.yaml
@@ -26,10 +26,6 @@ jobs:
     - uses: r-lib/actions/setup-r@v2
       with:
         r-version: '4.4.0'  # Specify the latest R version
-    - name: Install dependencies
-      run: |
-        Rscript -e "install.packages(c('devtools', 'roxygen2'))"
-        Rscript -e "devtools::install_deps(dependencies = TRUE)"
     - uses: ropensci-review-tools/pkgcheck-action@main
       with:
           summary-only: false
