### 2024-05-06 8509ebd
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@5446105bdeae38425a34e047b5bb33034bd39414 🚀
