### 2024-05-06 95c6e42
Update pkgcheck.yaml
