+++ .github/workflows/pkgcheck.yaml
@@ -14,6 +14,9 @@ on:
       - main
       - master
 
+env:
+  R_VERSION: '4.4.0'
+  
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
@@ -22,10 +25,6 @@ jobs:
     steps:
     - name: Check R version
       run: R --version
-    - uses: actions/checkout@v2
-    - uses: r-lib/actions/setup-r@v2
-      with:
-        r-version: '4.4.0'  # Specify the latest R version
     - uses: ropensci-review-tools/pkgcheck-action@main
       with:
           summary-only: false
