### 2024-05-06 39979af
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@95c6e42330c81af262e1aa74bb19d0c818dbb234 🚀
