### 2024-05-06 5bd0b43
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@847fa883a7d07310ec7d2a097d099a5c83fedf2f 🚀
