+++ .github/workflows/pkgcheck.yaml
@@ -15,14 +15,21 @@ on:
       - master
   
 jobs:
+  cpr:
+    name: job
+    runs-on: ubuntu-latest
+    steps:
+      - uses: r-lib/actions/setup-r@v2
+      - shell: Rscript {0}
+        run: |
+          if (!requireNamespace("MASS", quietly = TRUE))
+            install.packages("MASS")
+          library(MASS)
   pkgcheck:
     runs-on: ubuntu-latest
     permissions:
       issues: write
     steps:
-    - uses: r-lib/actions/setup-r@v2
-      with:
-        r-version: '4.3.0'
     - name: Check R version
       run: R --version
     - uses: ropensci-review-tools/pkgcheck-action@main
