### 2024-05-06 4dc1c38
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@fec2c4a180a7ea1052d366d9313855bd93abdef7 🚀
