### 2024-05-06 adbd31b
Update pkgcheck.yaml
