+++ .github/workflows/pkgcheck.yaml
@@ -18,6 +18,8 @@ jobs:
   cpr:
     name: job
     runs-on: ubuntu-latest
+    permissions:
+      issues: write
     steps:
       - uses: r-lib/actions/setup-r@v2
       - shell: Rscript {0}
@@ -25,13 +27,8 @@ jobs:
           if (!requireNamespace("MASS", quietly = TRUE))
             install.packages("MASS")
           library(MASS)
-  pkgcheck:
-    runs-on: ubuntu-latest
-    permissions:
-      issues: write
-    steps:
-    - name: Check R version
-      run: R --version
-    - uses: ropensci-review-tools/pkgcheck-action@main
-      with:
+      - name: Check R version
+        run: R --version
+      - uses: ropensci-review-tools/pkgcheck-action@main
+        with:
           summary-only: false
