### 2024-05-06 c480d49
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@adbd31b99545547246a8c0ec6521d501cc51ea41 🚀
