### 2024-05-06 c5adf62
Update pkgcheck.yaml
