+++ .github/workflows/pkgcheck.yaml
@@ -12,11 +12,10 @@ on:
   push:
     branches:
       - main
-      - master
   
 jobs:
   cpr:
-    name: job
+    name: pkgcheck
     runs-on: ubuntu-latest
     permissions:
       issues: write
@@ -24,9 +23,9 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
       - shell: Rscript {0}
         run: |
-          if (!requireNamespace("MASS", quietly = TRUE))
-            install.packages("MASS")
-          library(MASS)
+          if (!requireNamespace("sn", quietly = TRUE))
+            install.packages("sn")
+          library(sn)
       - name: Check R version
         run: R --version
       - uses: ropensci-review-tools/pkgcheck-action@main
