### 2024-05-06 b9b106b
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@c5adf6282587e1337727bba2832615642145d5e5 🚀
