### 2024-05-06 8398be1
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@32502b4399d87136cf60d26a7295aa8249993b07 🚀
