### 2024-05-10 4eb81e3
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@d5a6cc148616ba82df2aff482e7798bd3bfc5827 🚀
