### 2024-05-13 5f9f8d7
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@297d95fb2ebacf9dae1c7187b501e41e89df65db 🚀
