### 2024-05-13 b34cd46
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@6eb6aa5911e4f79effe079b12055b73098075229 🚀
