### 2024-05-13 fa38538
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@3f7430089b3ddac33b41679fdf1eac2d1b0038dc 🚀
