+++ vignettes/generate_rpkb.R
@@ -1,18 +1,0 @@
-## ----srr-tags, eval = FALSE, echo = FALSE-------------------------------------
-#  #' srr tags
-#  #'
-#  #'
-#  #' @srrstats {PD4.2} The results of the random sampling using the different
-#  #'                   distributions are compared graphically, by plotting on the
-#  #'                   sphere the generated observations.
-#  #'
-
-## -----------------------------------------------------------------------------
-library(QuadratiK)
-
-## -----------------------------------------------------------------------------
-mu <- c(0,0,1)
-d <- 3
-n <- 1000
-rho <- 0.8
-
