### 2024-05-13 1d41f3b
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@7bfdb79b98f0cfd39557fbd8bb83b4933501fcac 🚀
