### 2024-05-13 f3e8b5c
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@739598d6967e54eb8b8f5a81ca1d80fd351a60e2 🚀
