### 2024-05-14 dfab608
CRAN corrections
corrected arxiv doi and CRAN URL.
