+++ DESCRIPTION
@@ -16,11 +16,11 @@ Description: It includes test for multivariate normality, test for
     random generation of points from the Poisson kernel-based density and
     clustering algorithm for spherical data. For more information see
     Saraceno, G., Markatou, M., Mukhopadhyay, R., Golzy, M. (2024)
-    <arXiv:2402.02290>, Ding, Y., Markatou, M., Saraceno, G. (2023)
+    <doi:10.48550/arXiv.2402.02290>, Ding, Y., Markatou, M., Saraceno, G. (2023)
     <doi:10.5705/ss.202022.0347>, and Golzy, M., Markatou, M. (2020)
     <doi:10.1080/10618600.2020.1740713>.
 License: GPL (>= 3)
-URL: https://cran.r-project.org/web/packages/QuadratiK/index.html, https://github.com/giovsaraceno/QuadratiK-package, https://giovsaraceno.github.io/QuadratiK-package/
+URL: https://cran.r-project.org/package=QuadratiK, https://github.com/giovsaraceno/QuadratiK-package, https://giovsaraceno.github.io/QuadratiK-package/
 BugReports: https://github.com/giovsaraceno/QuadratiK-package/issues
 Depends: 
     R (>= 3.5.0)
