### 2024-05-14 5e840b2
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@dfab6088a467b872c417fb9e481ed5519e22e94e 🚀
