### 2024-05-17 2a43c8b
fix small issue
computation of the normality test statistics
