### 2024-05-17 65bed2d
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@2a43c8b061c6065d72776149ec35dc9c1057d0b1 🚀
