### 2024-05-17 27f05e4
updated variance computation
