+++ src/kernel_function.cpp
@@ -262,26 +262,31 @@ Eigen::VectorXd var_two(const Eigen::MatrixXd& Kcen,
  double est_var_Tr = 2 * n_factor * n_factor * pow(K_xx.array(),2).sum() +
     2 * m_factor * m_factor * pow(K_yy.array(),2).sum();
  
- double delta1 = 0.0;
- for (int i = 0; i < n; ++i) {
-    for (int j = 0; j < n; ++j) {
-       if (i != j) {  
-          for (int l = 0; l < m; ++l) {
-             delta1 += K_xx(i, j) * K_xy(i, l);
-          }
-       }
-    }
- }
- double delta2 = 0.0;
- for (int i = 0; i < m; ++i) {
-    for (int j = 0; j < m; ++j) {
-       if (i != j) {  
-          for (int l = 0; l < n; ++l) {
-             delta2 += K_yy(i, j) * K_xy(l, i);
-          }
-       }
-    }
- }
+ K_xx.diagonal().setZero();
+ double delta1 =  (K_xx.transpose() * K_xy).sum();
+ 
+  // double delta1 = 0.0;
+ // for (int i = 0; i < n; ++i) {
+ //    for (int j = 0; j < n; ++j) {
+ //       if (i != j) {  
+ //          for (int l = 0; l < m; ++l) {
+ //             delta1 += K_xx(i, j) * K_xy(i, l);
+ //          }
+ //       }
+ //    }
+ // }
+ K_yy.diagonal().setZero();
+ double delta2 =  (K_yy.transpose() * K_xy.transpose()).sum();
+ 
+ // for (int i = 0; i < m; ++i) {
+ //    for (int j = 0; j < m; ++j) {
+ //       if (i != j) {  
+ //          for (int l = 0; l < n; ++l) {
+ //             delta2 += K_yy(i, j) * K_xy(l, i);
+ //          }
+ //       }
+ //    }
+ // }
 
  est_varD -= 8 * n_factor * cross_factor * delta1;
  est_varD -= 8 * m_factor * cross_factor * delta2;
@@ -380,58 +385,97 @@ Eigen::VectorXd stat2sample(Eigen::MatrixXd& x_mat, Eigen::MatrixXd& y_mat, doub
 //' 
 //' @keywords internal
 // [[Rcpp::export]]
-Eigen::VectorXd var_k(const Eigen::MatrixXd& Kcen, const Eigen::VectorXd& sizes,
-                      const Eigen::VectorXd& cum_size) 
-{
-   int n = Kcen.rows();
+Eigen::VectorXd var_k(const Eigen::MatrixXd& Kcen, const Eigen::VectorXd& sizes, const Eigen::VectorXd& cum_size) {
    int K = sizes.size();
    
    Eigen::MatrixXd Kcen_copy = Kcen;
    Kcen_copy.diagonal().setZero();
    
-   double C1 = 0.0;
-   double C2 = 0.0;
-   double C3 = 0.0;
-   for (int l = 1; l <= K; ++l) {
+   double C1 = 0.0, C2 = 0.0, C3 = 0.0;
+   
+   for (int l = 0; l < K; ++l) {
+      double ni_factor = 1.0 / (sizes[l] * (sizes[l] - 1));
+      Eigen::MatrixXd K_ll = Kcen_copy.block(cum_size[l], cum_size[l], sizes[l], sizes[l]);
+      C1 += 2 * ni_factor * ni_factor * K_ll.array().square().sum();
       
-      double ni_factor = 1.0 / (sizes[l-1] * (sizes[l-1] - 1));
[…]
