### 2024-05-17 14f8c55
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@27f05e47f7af96d61cf07af058f98ef6500811be 🚀
