### 2024-05-17 5238039
compile package
- fix computation of the variance for the k-sample improving the computational time
- version 1.1.1
