+++ vignettes/TwoSample_test.Rmd
@@ -33,6 +33,7 @@ y_2 <- rmsn(n=n, xi=0, Omega = diag(d), alpha=rep(skewness_y,d))
 h <- 2
 set.seed(2468)
 two_test <- kb.test(x=x_2, y=y_2, h=h)
+two_test
 ```
 
 For the two-sample case, the ``summary`` function provides the results from the test and the descriptive statistics per variable and per group, as similarly described for the $k$-sample test. Additionally, it generates the qq-plots comparing the quantiles of the two groups for each variable. 
