### 2024-05-17 2fd76f2
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@523803917c2f1c5ea3fa01cd59fb9b2ffc44e5e3 🚀
