### 2024-05-22 3bed7b1
print kb.test
modify print and summary methods for kb.test
