### 2024-05-22 ab4dec8
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@3bed7b100d8acc4ad799b44d3f32733d6d4a8dcd 🚀
