### 2024-05-31 giovsaraceno/QuadratiK-package#5 @mpadge
Unable to load package: Invalid ELF header
@giovsaraceno @rmj3197 Thanks for submitting your package to [rOpenSci review](https://github.com/ropensci/software-review/issues/632). Our automated checking system failed because your package fails to load with this error:
```
Errorr in dyn.load(dll_copy_file) :
  unable to load shared object '/tmp/RtmpTkQfgu/pkgload6516c1c5f36/QuadratiK.so':
  /tmp/RtmpTkQfgu/pkgload6516c1c5f36/QuadratiK.so: invalid ELF header
In addition: Warning messages:
1: In rgl.init(initValue, onlyNULL) : RGL: unable to open X11 display
2: 'rgl.init' failed, running with 'rgl.useNULL = TRUE'.
```
This can be easily reproduced by cloning and running docker container `mpadge/pkgcheck`, cloning your repo in there, and then running `devtools::load_all(<path/to/QuadritiK-package>)`.
