### 2024-05-31 #632 @mpadge
@jooolia The automated checks failed because of issue linked to above. @giovsaraceno When you've fixed this issue and confirmed that pkgcheck workflows once again succeed in your repo, please call `@ropensci-review-bot check package` here to run checks again. Thanks
