### 2024-05-31 dc3d2a8
Add data sets
-addition of two data sets
- show method for two and k sample tests
- two-sample qqplots with different sample sizes
