+++ man/QuadratiK-package.Rd
@@ -47,7 +47,7 @@ the Sphere: Convergence Properties, Identifiability, and a Method of
 Sampling, Journal of Computational and Graphical Statistics, 29:4, 758-770,
 DOI: 10.1080/10618600.2020.1740713.
 
-Markatou M, Saraceno G, Chen Y (2023). “Two- and k-Sample Tests Based on
+Markatou M, Saraceno G, Chen Y (2024). “Two- and k-Sample Tests Based on
 Quadratic Distances.” Manuscript, (Department of Biostatistics,
 University at Buffalo).
 }
