+++ vignettes/kSample_test.Rmd
@@ -1,6 +1,5 @@
 ---
 title: "k-sample test"
-author: "Giovanni Saraceno"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{k-sample test}
@@ -58,5 +57,3 @@ h_k <- select_h(x=x, y=y, alternative="skewness")
 h_k$h_sel
 ```
 The figure generated by the ``select_h`` function on the result of the selection of $h$ algorithm for the $k$-sample data set displays the obtained power versus the considered $h$, for each value of skewness alternative $\delta$ considered.
-
-As it is possible to see from the figure, when the alternative distribution $F_\delta$ with $\delta=0.2$ is considered, there are no values of $h$ which achieve power greater than or equal to 0.5. Then, the second value of $\delta=0.3$ is take into account and $h=1.6$ is chosen as optimal value since it is the smallest value with power greater than 0.5. Additionally, it gives a possible set of values with high power performance.

[…]
