### 2024-05-31 23d3b18
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@dc3d2a88fcc5c3a44194ab9f5755788c4f41392c 🚀
