### 2024-05-31 543ab97
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@245381241c2679f3b2bc50c6d8535da2038e9b0f 🚀
