### 2024-05-31 072f953
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@b438a339cb8c12350e6cbf1cb8bb148913d49c5b 🚀
