+++ R/QuadratiK-package.R
@@ -1,4 +1,3 @@
-#' @docType package
 #' @name QuadratiK-package
 #' @title Collection of Methods Constructed using the Kernel-Based Quadratic 
 #'          Distances
@@ -54,4 +53,5 @@
 #' @srrstats {G3.0} in all the codes floating point numbers are not compared
 #' @srrstats {G5.7, G5.9} Verified through the simulation studies in the 
 #'                        indicated references. 
-NULL
+#' @keywords internal 
+"_PACKAGE"
