### 2024-05-31 a9da31b
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@6be7409a18441e7c037c7571955fbf1198c4c04a 🚀
