+++ NAMESPACE
@@ -31,7 +31,7 @@ importFrom(Tinflex,Tinflex.sample)
 importFrom(Tinflex,Tinflex.setup.C)
 importFrom(cluster,silhouette)
 importFrom(clusterRepro,IGP.clusterRepro)
-importFrom(ggpp,annotate)
+importFrom(ggpp,geom_table_npc)
 importFrom(ggpubr,ggarrange)
 importFrom(grDevices,colorRampPalette)
 importFrom(mclust,adjustedRandIndex)
