### 2024-06-02 a354b11
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@b5c45c22193324f0d0fcc92764a47a0e048b9937 🚀
