+++ DESCRIPTION
@@ -29,8 +29,8 @@ Imports:
     clusterRepro,
     doParallel,
     foreach,
-    ggplot2,
     ggpp,
+    ggplot2,
     ggpubr,
     mclust,
     methods,
