### 2024-06-02 d6f34ba
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@bc0c1fff4998d40db684ad79abf2ecffe87dfee8 🚀
