### 2024-06-03 9f9e6d0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@577253bcf04bb66dcae598b87b8c1532b5b5b4d6 🚀
