### 2024-06-04 d6b6bf4
documentation
changes in text for the manual and vignettes
