+++ R/QuadratiK-package.R
@@ -2,7 +2,7 @@
 #' @title Collection of Methods Constructed using the Kernel-Based Quadratic 
 #'          Distances
 #' @description It is implemented in R and Python, providing a comprehensive set
-#' of goodness-of-fit tests and clustering technique using kernel-based 
+#' of goodness-of-fit tests and a clustering technique using kernel-based 
 #' quadratic distances. This framework aims to bridge the gap between the 
 #' statistical and machine learning literature. It includes:
 #' \itemize{
@@ -53,5 +53,5 @@
 #' @srrstats {G3.0} in all the codes floating point numbers are not compared
 #' @srrstats {G5.7, G5.9} Verified through the simulation studies in the 
 #'                        indicated references. 
-#' @keywords internal 
+#' 
 "_PACKAGE"
