+++ vignettes/TwoSample_test.Rmd
@@ -14,8 +14,10 @@ vignette: >
 #' 
 #' @srrstats {G1.5} two-sample test example in the associated paper
 ```
-The two-sample test can be additionally performed by providing the two sample to be compared as ``x`` and ``y``. We generate the sample $y = (y_1, \ldots,y_n)$ from a skew-normal distribution $SN_d(0,I_d, \lambda)$, where $d=4$, $n=200$ and $\lambda= (0.5,\ldots,0.5)$.  
-
+The two-sample test can be additionally performed by providing the two samples to be compared as ``x`` and ``y``. We generate the sample $y = (y_1, \ldots,y_n)$ from a skew-normal distribution $SN_d(0,I_d, \lambda)$, where $d=4$, $n=200$ and $\lambda= (0.5,\ldots,0.5)$.  
+```{r setup, include=FALSE} 
+knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
+```
 ```{r message=FALSE}
 library(sn)
 library(mvtnorm)
@@ -48,7 +50,7 @@ The figure automatically generated by the ``summary`` function on the result of 
 
 #### Select h
 
-If a value of $h$ is not provided, the function automatically perform the function ``select_h``.
+If a value of $h$ is not provided, the function automatically performs the function ``select_h``.
 ```{r}
 # two_test_h <- kb.test(x=x_2, y=y_2)
 ```
@@ -63,5 +65,5 @@ h_test2 <- select_h(x=x_2, y=y_2, alternative="skewness")
 h_test2$h_sel
 ```
 
-The figure generated by the ``select_h`` function on the result of the selection of $h$ algorithm on the two-sample data set displays the obtained power versus the considered $h$, for the alternatives $\delta$.
+The function``select_h`` will also generate a figure which displays the obtained power versus the considered $h$, for the alternatives $\delta$.
 
