+++ vignettes/kSample_test.Rmd
@@ -35,7 +35,7 @@ k_test <- kb.test(x=x, y=y, h=h)
 k_test
 ```
 
-When the $k$-sample test is performed, the ``summary`` method on the ``kb.test`` object returns the results of the tests together with the standard descriptive statistics for each variable computed, overall and with respect to the provided groups. 
+When the $k$-sample test is performed, the ``summary`` method on the ``kb.test`` object returns the results of the tests together with the standard descriptive statistics for each variable computed, overall, and with respect to the provided groups. 
 ```{r}
 summary_ktest <- summary(k_test)
 summary_ktest$summary_tables
@@ -43,7 +43,7 @@ summary_ktest$summary_tables
 
 #### Selection of h
 
-If a value of $h$ is not provided, the function automatically perform the function ``select_h``.
+If a value of $h$ is not provided, the function automatically performs the function ``select_h``.
 ```{r}
 #k_test_h <- kb.test(x=x, y=y)
 ```
@@ -56,4 +56,4 @@ h_k <- select_h(x=x, y=y, alternative="skewness")
 ```{r eval=FALSE}
 h_k$h_sel
 ```
-The figure generated by the ``select_h`` function on the result of the selection of $h$ algorithm for the $k$-sample data set displays the obtained power versus the considered $h$, for each value of skewness alternative $\delta$ considered.
+The ``select_h`` function will also generate a figure displaying the obtained power versus the considered $h$, for each value of skewness alternative $\delta$ considered.
