### 2024-06-04 d26aa23
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@d6b6bf4747f58a49d75ced4c334190960b1237bb 🚀
