### 2024-06-04 giovsaraceno/QuadratiK-package#5 @giovsaraceno
We have fixed the reported error. Now the pkgcheck is running correctly. Thanks.
