### 2024-06-17 21df45c
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@12a798410335a68f08b4a459d96e1636139b57e6 🚀
