### 2024-06-18 f4c5398
Update NEWS.md
