### 2024-06-18 6a6791c
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@00b3338c6b9ee84b74e0b7055b46b0b3c7d7346e 🚀
