### 2024-06-18 77e9748
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@f4c5398b8c776a88206c1499cc1485a8b264c466 🚀
