### 2024-06-24 #632 @mpadge
@emitanaka Sorry about that. The issue template at the very top had been modified, including removing the "editor" field needed by the bot to identify you. I've reinstated everything now, so should be okay.

@giovsaraceno There are a couple of fields which still need to be filled in. Can you please edit the top of the initial issue text at the top and fill out:

- [ ] Version submitted
- [ ] "Badge grade" - choose one and delete all others

Thanks!
