### 2024-06-24 #632 @giovsaraceno
@mpadge I have modified the initial issue text by additing the version submitted and choosing the badge grade. Please let us know if anything else is needed.
Thanks!
