### 2024-06-24 #632 @giovsaraceno
> Please add this badge to the README of your package repository:
>
> ```
> ```
>
> Furthermore, if your package does not have a NEWS.md file yet, please create one to capture the changes made during the review process. See https://devguide.ropensci.org/releasing.html#news

We have added the provided badge into the README file and added the NEWS.md file into the package.
