### 2024-06-28 #632 @kasselhingee
- I have no relationship past or present with the package authors.

---

I'm only partially familiar with the area of spherical data. Both your package's tests of uniformity and clustering sound useful. The `kb.test()` tests sound complicated but I'm sure with more explanation their use will become clear.

I found it really hard to understand your package initially. That was because most of the documentation ignores [G1.3](https://stats-devguide.ropensci.org/standards.html#statistical-terminology) on explaining statistical terms. For example, I jumped into the `kb.test()` part of the package which was really confusing until I found an arXiv document on QuadratiK that had more information.
Although the other parts were less confusing to me, they still didn't explain themselves well, often stating that the function performs "the [a bespoke method by authors]" without explanation of the method.
Because of this, currently your package feels only usable by expert statisticians who have read your papers and want to try out your methods.

I would love it if your readme described the major benefits with more detail. For example, that `pk.test()` performs much better than competitors when the alternative is multimodal.
Your answers to rOpenSci [here](https://github.com/ropensci/software-review/issues/632#issue-2185032498) state that novel and unique kernel-based methods are used, but you don't say what is good about them. Also, please clarify if the kernel in the 'Poisson kernel-based densities' different from the kernel in your 'kernel-based quadratic distances'.

More on following [G1.3](https://stats-devguide.ropensci.org/standards.html#statistical-terminology). There are many unexplained terms I've never heard of before, a few used differently to expectation, and others used vaguely. At times it felt like these were all the terms!

  + An intro to PKB distributions would be nice.
  + And exact meaning of parameters in `rpkb()`.

### Package-Level Comments
#### Major Functionality Claims
The following appear to be met from the documentation and function results:

+ A test of uniformity on the sphere for multimodal alternatives
+ Clustering algorithms
+ Equal-distribution tests (this is my current guess of what the k-sample tests are)
+ Goodness of fit to an estimated PKB distribution?
+ Some graphics help

However, I haven't checked if they are implemented correctly and the package lacks tests to confirm results in simple situations.

These claims seem *unmet* to me:

+ I can't see the bridge between statistics and machine learning claimed in the README
+ I haven't got the scatterplots of clustering results to work
+ I can't see the Python version
+ In answering some rOpenSci questions you mentioned a dashboard application. I completely missed it! Where is it?

#### Statement of need in README
+ Statement of need lacks precision. When should I use your methods vs existing methods?
+ The majority of methods are based on statistical methods published with peer-review by the authors, so I haven't checked the novelty or importance of the methods myself.
An exception is the 'k-sample goodness of fit tests', which are in a manuscript I couldn't find. Not sure what they are exactly, or how good they are compared to existing methods.
+ Two and k-sample tests often refer to testing equality between groups (or similar).
  + Please be more clear in the README what the hypotheses are in these tests.
  + I'm especially confused by a k-sample test of goodness of fit to probability distribution. Without reading your arXiv document it sounded like it would test the fit of each sample to the model probability distributions.
  + What model does the single-sample goodness of fit test?
  + Specify what class of probability distributions (PKB?)
+ Target audience is unclear in README. Currently I'm thinking expert statisticians and machine learners who have read your papers/manuscripts, but I see your answers to rOpenSci questions hope for more general use.

+ Doesn't have installation instructions in README for the development version or CRAN version (but should be very easy to add: I did it with `devtools::install_github("https://github.com/giovsaraceno/QuadratiK-package/tree/master")`)

#### Dependencies
+ Why is ggplot2, rlecuyer, doParallel, foreach, imported but aren't called according to the automatic check?
+ Package has lots of dependencies (imports) which could make it hard to maintain. rrcov, mvtnorm, movMF, moments, mclust, clusterRepro, cluster, Tinflex, ggpp, ggpubr all seem to be only called once. Are they all crucial? Could they be optional extra functionality?

### Parts of the Package
+ Seem to cover the main uses of the package
+ One vignette can't be run locally
+ Overall they appear to be written for people already familiar with the authors' papers or the other vignettes.

+ `wireless_clustering` vignette
  + Please justify this conversion to the sphere more. "Given that the Wi-Fi signal strength takes values in a limited range, it is appropriate to consider the spherically transformed observations, by L2 normalization, and consequently perform the clustering algorithm on the 7-dimensional sphere." On the face of it a sphere seems in appropriate, why doesn't the total length of a vector in the space matter?
  + Typo: "it returns an object with IGP the InGroup Proportion (IGP), and metrics a table of computed evaluation measures" --> remove 'IGP the'.
  + A chunk is visible but isn't evaluated, which confused me a lot, because the function inside the chunk called `validation()` doesn't exist. I'm guessing it is now `pkbc_validation()`.
  + What is Average Silhouette Width etc. from `pkbc_validation()`?
  + The paragraph on plotting is unclear which object the plot method is for. A `plot(res_pk)` would be nice, with an example plot showing why K=4 looks appropriate.

+ k-sample test vignette
  + "Recall that the computed test statistics correspond to the omnibus tests." Recall from where? And what are the omnibus tests?
  + The commented-out line in the code chunk "#k_test_h <- kb.test(x=x, y=y)" confused me. I think you mean it to be an example of automatic use of `select_h()`?
  + `h_k <- select_h(x=x, y=y, alternative="skewness")` takes a really long time on my machine, would be good to mention in the help for `select_h()` that it takes a long time.

+ I quickly scanned and checked the remaining vignettes could render, but didn't run the code myself.

#### Function help
+ I suspect help assumes user has read your references e.g.,
  + For `pkbc_validation()`, how do I interpret all the measures?
  + For `kb.test()` see my comments about k-sample tests in the README.

+ `kb.test()`
  + What happens if the reference distribution parameters are `NULL`?
  + I googled for your 2024 manuscript but couldn't find it, so not sure what the tests are.
  + There are many aspects not seen documented in `kb.test()`. What are the U-statistics, what is Vn, and a V-statistic etc?
  + The function clearly does a lot of things, but what are those things? All you've said is "the kernel-based quadratic distance tests using the Gaussian kernel with bandwidth parameter h". Please describe what they are.

+ `kb.test` class
  + Great to have this class documented explicitly.
  + Link to `kb.test()` and vice versa to explain the objects.
  + The class and the function are so closely linked so consider documenting them in the same .Rd file. But it is okay for them to be separate.

+ `pk.test()` and `pk.test-class`
  + What happens if `rho` is NULL?
  + Would be great to say somewhere when `pk.test()` is advised - from the referenced paper it is when the alternative is multimodal.

+ `pkbc()`
  + Written like the user is familiar with the referenced paper

+ `pkbc_validation()`
  + I'm guessing the columns of `pkbc_validation(res)$metrics` are the different cluster numbers in `res`?

+ `plot.pkbc()`
  + For some reason on my system after installing from CRAN: `?plot.pkbc` and `help(plot.pkbc)` both find nothing. But, for example `help(plot.TinflexC)` gets me to the appropriate help. Do you know why it isn't working? The manual for `plot.pkbc` is crucial to understanding the plots so should be accessible from the console.
  + Can you please make it so that it can be used non-interactively?
     + This is crucial for reproducibility for anyone using this function.
  + Please document what the options mean. In the `wireless_clustering` vignette `plot(res_pk)` choosing scatter plot then 4 clusters to display didn't make a plot at all! Is this a bug, or what it should be doing? And surely it means displaying the 4-cluster model, rather than say displaying the first 4 clusters of the 6-cluster model?
  + It would be nice if the elbow plot panels made clear which metric they are using, but I know that sort of thing can be hard to program. I'm hoping the order corresponds to the order in the help (left = Euclidean, right = cosine similarity).

+ `select_h()`
  + Can take a really long time - worth warning the user

+ `wine` data set
  + Source seems to be missing some spaces

+ `wireless` data set
  + Source seems to be missing some spaces

#### Examples
I get warnings from `geom_table_npc()` when I run the examples. I'm using version 0.5.7 of package ggpp.

### Test Coverage
+ Could you test that `pk.test()` rejects uniformity correctly?
+ You could test rpkb against a known approximation with preexisting simulation methods (e.g. vMF, kent etc?) (currently you only test the mean)
+ Test of `pkbc()` doesn't seem to check that it gets the clustering correct in a simple situation
+ Likewise 'stats_clusters()` and `pkbc_validation()` (these all seem to test that the structure/class of the output is correct, but not the actual values of the output)
+ Considering your value of the graphical methods (and that I couldn't get one to work), testing that they work would be good!

### Packaging guidelines
+ README should include a brief demonstration
+ No citation instructions yet
+ Guidelines suggest a general vignette to introduce the package
+ Vignettes are missing citations
+ There are places where cross-links would be appropriate to add (e.g. between pb.test help and pb.test-class, link to plot.kpbc help from kbpc and vice-versa).
+ Do you have a pkgdown website yet?

### More
+ G1.6 suggests a demo of `pk.test()` on multimodal data vs another uniform testing method
+ I can see a few spelling errors (e.g. silhouette) according to `devtools::spell_check()`
