### 2024-07-14 #632 @kasselhingee
Hi @giovsaraceno , glad it was helpful.

+ optional extra functionality: I haven't done this much myself, but I would put these packages as 'Suggested' in the DESCRIPTION file and inside `pkbc_validation()` check for mclust etc (using `require(pkgname)` say). I've can imagine two situations then (1) ARI, IGP and ASW are core to `pkbc_validation()`, in which case I've seen other people's packages ask me if I want to install the packages (not sure how they did it) OR (2) they aren't core to  `pkbc_validation()` and you could just print a `message()` if the packages aren't available. In either case you'll have to make sure the help for `pkbc_validation()` explains the behaviour.

+ Test coverage for `rpkb()`: I got the impression that certain parameter settings for the Poisson kernel-based distribution can closely approximate a vMF or Kent etc. I suggest checking that the results of `rpkb()` are very similar to the results for existing vMF and Kent etc simulators in at least one of these situations. (And if the situation is close to a typical use case for `rpkb()` that will be an even better test!).
