### 2024-08-27 093604f
Review 1
Changes on documentation following the suggestions and comment from Reviewer 1 rOpenSci
