+++ R/clustering_functions.R
@@ -1,28 +1,66 @@
 #'
 #' Poisson kernel-based clustering on the sphere
 #'
-#' The function \code{pkbc} performs the Poisson kernel-based clustering 
-#' algorithm on the sphere based on the Poisson kernel-based densities.
-#'
+#' @description
+#' The function \code{pkbc()} performs the Poisson kernel-based clustering 
+#' algorithm on the sphere proposed by Golzy and Markatou (2020). 
+#' The proposed algorithm is based on a mixture, with \eqn{M} components, of 
+#' Poisson kernel-based densities on the hypersphere \eqn{\mathcal{S}^{d-1}}
+#' given by
+#' \deqn{f(x|\Theta) = \sum_{j=1}^M \alpha_j f_j(x|\rho_j, \mu_j)}
+#' where \eqn{\alpha_j}'s are the mixing proportions and \eqn{f_j(x|\rho_j, 
+#' \mu_j)}'s denote the probability density function of a \eqn{d}-variate 
+#' Poisson kernel-based density given as
+#' \deqn{f(\mathbf{x}|\rho, \mathbf{\mu}) = \frac{1-\rho^2}{\omega_d 
+#' ||\mathbf{x} - \rho \mathbf{\mu}||^d}.}
+#' The parameters \eqn{\alpha_j, \mu_j, \rho_j} are estimated through a 
+#' iterative reweighted EM algorithm. \cr
+#' The proposed clustering algorithm exhibits excellent results when 
+#' (1) the clusters are not well separated; (2) the data points
+#' are fairly well concentrated around the vectors \eqn{\mu_j} of each cluster;
+#' (3) the percentage of noise in the data increases.
+#' 
 #' @param dat Data matrix or data.frame of data points on the sphere to be 
-#'            clustered. The observations in dat are normalized to ensure that
-#'            they lie on the d-simensional sphere. Note that d > 1.
+#'            clustered. The observations in \code{dat} are normalized to ensure
+#'            that they lie on the \eqn{d}-dimensional sphere. Note that 
+#'            \eqn{d > 1}.
 #' @param nClust Number of clusters. It can be a single value or a numeric 
 #'               vector.
 #' @param maxIter The maximum number of iterations before a run is terminated.
 #' @param stoppingRule String describing the stopping rule to be used within 
-#'                     each run. Currently must be either: 
-#'                \code{'max'} (until the change in the log-likelihood is less 
-#'                than a given threshold (1e-7)), 
-#'                \code{'membership'} (until the membership is unchanged), or 
-#'                \code{'loglik'} (based on a maximum number of iterations).
+#'                     each run. Currently must be either \code{'max'}, 
+#'                     \code{'membership'}, or \code{'loglik'}.
 #' @param initMethod String describing the initialization method to be used.
 #'                   Currently must be \code{'sampleData'}.
-#' @param numInit Number of initializations.
+#' @param numInit Number of initialization.
 #'
-#' @details The function estimates the parameter of a mixture of Poisson
-#' kernel-based densities. The obtained estimates are used for assigning final 
-#' memberships, identifying the \code{nClust} clusters.
+#' @details 
+#' We set all concentration parameters equal to 0.5 and all mixing proportions
+#' to be equal. \cr
+#' The initialization method \code{'sampleData'} indicates that observation 
+#' points are randomly chosen as initializers of the centroids \eqn{\mu_j}.
+#' This random starts strategy has a chance of not obtaining initial 
[…]
