+++ R/classes.R
@@ -6,7 +6,7 @@
 #' quadratic distance tests. This includes the normality test, the two-sample 
 #' test statistics and the k-sample tests.
 #'
-#' @slot method String indicating the normal kernel-based quadratic distance 
+#' @slot method String indicating the kernel-based quadratic distance 
 #'              test performed.
 #' @slot Un The value of the test U-statistics.
 #' @slot Vn The value of the test V-statistic.
@@ -63,8 +63,7 @@ setClass("kb.test",
 #' @description A class to represent the results of Poisson kernel-based 
 #'              quadratic distance tests for Uniformity on the sphere.
 #'
-#' @slot method The method used for the test ("Poisson Kernel-based quadratic 
-#'              distance test of Uniformity on the Sphere").
+#' @slot method Description of the test.
 #' @slot x Matrix of data
 #' @slot Un The value of the U-statistic.
 #' @slot CV_Un The critical value for Un computed through replications.
@@ -79,6 +78,8 @@ setClass("kb.test",
 #' @slot B Number of replications.
 #' @slot var_Un exact variance of the kernel-based U-statistic.
 #'
+#' @seealso [pk.test()] 
+#' 
 #' @examples
 #' # create a pk.test object
 #' d=3
@@ -116,7 +117,7 @@ setClass("pk.test",
 #'             each value of possible number of clusters considered.
 #' @slot input List of input data
 #'
-#' @details See the function \code{pkbc} for more details.
+#' @seealso [pkbc()] for more details.
 #' 
 #' @examples 
 #' data("wireless")
