+++ R/critical_value.R
@@ -23,9 +23,10 @@
 #' @return the critical value for the specified method and significance level.
 #' 
 #' @references
-#' Markatou Marianthi, Saraceno Giovanni, Chen Yang (2023). “Two- and k-Sample 
-#' Tests Based on Quadratic Distances.” Manuscript, (Department of 
-#' Biostatistics, University at Buffalo)
+#' Markatou Marianthi & Saraceno Giovanni (2024). “A Unified Framework for
+#' Multivariate Two- and k-Sample Kernel-based Quadratic Distance 
+#' Goodness-of-Fit Tests.” 
+#' https://doi.org/10.48550/arXiv.2407.16374
 #'
 #'
 #' @useDynLib QuadratiK
