### 2024-08-28 1bf0f00
documentation
Improved documenation and presentation of vignettes
