+++ R/QuadratiK-package.R
@@ -1,33 +1,65 @@
 #' @name QuadratiK-package
 #' @title Collection of Methods Constructed using the Kernel-Based Quadratic 
 #'          Distances
-#' @description It is implemented in R and Python, providing a comprehensive set
-#' of goodness-of-fit tests and a clustering technique using kernel-based 
-#' quadratic distances. This framework aims to bridge the gap between the 
-#' statistical and machine learning literature. It includes:
-#' \itemize{
-#'   \item \bold{Goodness-of-Fit Tests}: The software implements one, two, and 
-#'  k-sample tests for goodness of fit, offering an efficient and mathematically
-#'   sound way to assess the fit of probability distributions. Expanded 
-#'   capabilities include supporting tests for uniformity on the d-dimensional 
-#'   Sphere based on Poisson kernel densities.
-#'   \item \bold{Clustering Algorithm for Spherical Data}: the package 
-#'   incorporates a unique clustering algorithm specifically tailored for 
-#'   spherical data. This algorithm leverages a mixture of Poisson kernel-based
-#'   densities on the Sphere, enabling effective clustering of spherical data or
-#'   data that has been spherically transformed. The package also provides the 
-#'   functions for density evaluation and random sampling from the Poisson 
-#'   kernel-based distribution.
-#'   \item \bold{Additional Features}: Alongside these functionalities, the 
-#'   software includes additional graphical functions, aiding users in 
-#'   validating and representing the cluster results as well as enhancing the 
-#'   interpretability and usability of the analysis.
-#' }
+#' @description 
+#' Collection of Methods Constructed using the Kernel-Based Quadratic Distances
+#' 
+#' `QuadratiK` provides the first implementation, in R and Python, of a
+#' comprehensive set of goodness-of-fit tests and a clustering technique for
+#' spherical data using kernel-based quadratic distances. It includes:
+#' -   **Goodness-of-Fit Tests**: The software implements one, two, and 
+#' *k*-sample tests for goodness of fit, offering an efficient and 
+#' mathematically sound way to assess the fit of probability distributions.
+#' Expanded capabilities include supporting tests for uniformity on the 
+#' *d*-dimensional Sphere based on Poisson kernel densities. Our tests are
+#' particularly useful for large, high dimensional data sets where the 
+#' assessment of fit of probability models is of interest. Specifically, we 
+#' offer tests for normality, as well as two- and *k*-sample tests, where 
+#' testing equality of two or more distributions is of interest, that is 
+#' \eqn{H_0: F_1 = F_2} and \eqn{H_0: F_1 = \ldots = F_k} respectively.
+#' The proposed tests perform well in terms of level and power for contiguous
+#'  alternatives, heavy tailed distributions and in higher dimensions.
+#' -   **Poisson kernel-based distribution (PKBD)**: the package offers 
+#' functions for computing the density value and for generating random samples 
+#' from a PKBD. The Poisson kernel-based densities are based on the normalized 
+#' Poisson kernel and are defined on the $d$-dimensional unit sphere. Given a 
+#' vector \eqn{\mu \in \mathcal{S}^{d-1}}, and a parameter \eqn{\rho} such that 
+#' \eqn{0 < \rho < 1}, the probability density function of a \eqn{d}-variate 
+#' Poisson kernel-based density is defined by: \deqn{f(\mathbf{x}|\rho,
[…]
