+++ R/wireless.R
@@ -29,7 +29,7 @@
 #'
 #' @source
 #' Bhatt, Rajen (2017). Wireless Indoor Localization. 
-#' UCI Machine Learning Repository. 
+#' UCI Machine Learning Repository. \cr
 #' https://doi.org/10.24432/C51880.
 #'
 #' @references

[…]
