+++ .github/workflows/pkgcheck.yaml
@@ -13,6 +13,7 @@ on:
     branches:
       - main
       - master
+      - review-rOpenSci
 
 jobs:
   pkgcheck:
