+++ .github/workflows/R-CMD-check.yaml
@@ -6,9 +6,9 @@
 # usethis::use_github_action("check-standard") will install it.
 on:
   push:
-    branches: [main, master]
+    branches: [main, master, review-rOpenSci]
   pull_request:
-    branches: [main, master]
+    branches: [main, master, review-rOpenSci]
 
 name: R-CMD-check
 
