### 2024-09-02 61526d5
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@2c80fe0d1030f7d559ebf4484647267ddc6c388a 🚀
