### 2024-09-02 c6cd53d
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@c635a7724141d46b0fffae5e7b66ff7bbc07f0b9 🚀
