+++ tests/testthat/test-rpkb.R
@@ -96,3 +96,30 @@ test_that("Random Generation from PKBD works", {
    
    
 })
+
+test_that("Random Generation from PKBD compared to wrapped Cauchy", {
+   
+   require(circular)
+   
+   # Parameters 
+   n <- 1000  
+   location <- circular(0)  
+   rho <- 0.5  
+   # Generate data from wrapped cauchy
+   wc <- rvonmises(n, mu = location, kappa = -log(rho))
+   # Generate data from pkbd 
+   pkbd <- rpkb(n, mu = c(1, 0), rho = 0.5)$x
+   
+   # Convert Cartesian coordinates to angles for comparison
+   pkbd_angles <- atan2(pkbd[,2], pkbd[,1]) %% (2*pi)
+   # Convert PKBD angles to circular data
+   pkbd_circular <- circular(pkbd_angles)
+   # Convert to circular objects for statistical testing
+   wc_circular <- circular(wc)
+   
+   # Perform Watson-Wheeler test 
+   # (common for comparing two circular distributions)
+   ww_test <- watson.two.test(wc_circular, pkbd_circular)
+   
+   
+})
