+++ tests/testthat/test-pk.test.R
@@ -63,15 +63,24 @@ test_that("Functionality with valid inputs", {
    
    size <- 100
    d <- 3
+   
+   # Test does not reject uniformity
    set.seed(123)
    x_sp <- sample_hypersphere(d, n_points=size)
-   
    unif_test <- pk.test(x_sp,rho=0.8)
    
    expect_s4_class(unif_test, "pk.test")
    expect_true(is.numeric(unif_test@Un))
    expect_true(is.numeric(unif_test@Vn))
    expect_false(unif_test@H0_Un)
+   
+   # Test does not reject uniformity
+   set.seed(123)
+   x_sp <- rpkb(n = size, mu = c(1,0,0), rho = 0.9)$x
+   unif_test <- pk.test(x_sp,rho=0.9)
+   
+   expect_s4_class(unif_test, "pk.test")
+   expect_true(unif_test@H0_Un)
    
    # test show method
    output <- capture.output(show(unif_test))

[…]
