### 2024-09-02 b3d94da
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@59bb992336bc817eaa625f048d10cf910d7900ad 🚀
