### 2024-09-02 504f44e
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@44177e8ff66780e81baebcfc33910ce4bd0e919f 🚀
