### 2024-09-04 08bf146
extra functionality
move packages into "Suggest" section of DESCRIPTION file
