### 2024-09-04 c85df09
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@b45bccd52080696f6f1d5b34444291d06b20216a 🚀
