+++ pkgdown.yml
@@ -7,4 +7,4 @@ articles:
   TwoSample_test: TwoSample_test.html
   uniformity: uniformity.html
   wireless_clustering: wireless_clustering.html
-last_built: 2024-09-02T14:12Z
+last_built: 2024-09-04T13:12Z
