### 2024-09-04 3f0dd8c
error in vignettes
add paper in DESCRIPTION file
