+++ DESCRIPTION
@@ -16,7 +16,8 @@ Description: It includes test for multivariate normality, test for
     random generation of points from the Poisson kernel-based density and
     clustering algorithm for spherical data. For more information see
     Saraceno, G., Markatou, M., Mukhopadhyay, R., Golzy, M. (2024)
-    <doi:10.48550/arXiv.2402.02290>, Ding, Y., Markatou, M., Saraceno, G. (2023)
+    <doi:10.48550/arXiv.2402.02290>, Markatou, M., Saraceno, G. (2024) 
+    <doi:10.48550/arXiv.2407.16374>, Ding, Y., Markatou, M., Saraceno, G. (2023)
     <doi:10.5705/ss.202022.0347>, and Golzy, M., Markatou, M. (2020)
     <doi:10.1080/10618600.2020.1740713>.
 License: GPL (>= 3)
