### 2024-09-04 669c72b
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@3f0dd8ce9c8ccfef3c935a685df8f4e9abfe099f 🚀
