+++ DESCRIPTION
@@ -52,7 +52,8 @@ Suggests:
     clusterRepro,
     mclust,
     Tinflex,
-    movMF
+    movMF,
+    devtools
 LinkingTo: 
     Rcpp,
     RcppEigen
