### 2024-09-04 d478c78
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@e4531887be4eb2b3e8c45964179e7c68fc86cb7e 🚀
