### 2024-09-04 0c1d869
error vignettes
