### 2024-09-04 9ed0a34
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@0c1d8694996887321814542b21a8bd6078f9db0b 🚀
