### 2024-09-07 badf613
vignettes
add vignettes to .Rbuildignore (TwoSample_test and kSample_test) since the function select_h require that the package is already installed
