+++ NAMESPACE
@@ -23,10 +23,11 @@ exportMethods(summary)
 import(Rcpp)
 import(RcppEigen)
 import(doParallel)
-import(foreach)
 import(ggplot2)
 import(rlecuyer)
 importFrom(Rcpp,sourceCpp)
+importFrom(foreach,"%dopar%")
+importFrom(foreach,foreach)
 importFrom(ggpubr,ggarrange)
 importFrom(grDevices,rainbow)
 importFrom(graphics,legend)
