### 2024-09-07 7b6beb6
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@badf61380a3b4962401740bbe15f4de721ac8c88 🚀
