+++ R/pk.test.R
@@ -257,18 +257,6 @@ setMethod("summary", "pk.test", function(object) {
       
       stats[[i]] <- stats_step
       
-      # pl_stat <- ggplot() +
-      # geom_table_npc(data = data.frame(Stat = rownames(stats_step), stats_step),
-      #                aes(npcx = 0.5, npcy = 0.5, 
-      #       label = list(data.frame(Stat = rownames(stats_step), stats_step))),
-      #                   hjust = 0.5, vjust = 0.5) +
-      #   #annotate('table', x = 0.5, y = 0.5, 
-      #            #   label = data.frame(Stat = rownames(stats_step),stats_step),
-      #               # hjust = 0.5, vjust = 0.5) +
-      #    theme_void() +
-      #    ggtitle("")+
-      #    scale_color_brewer(palette='Set1')
-      # 
       plot_list[[length(plot_list) + 1]] <- list(pl)
       
       
