### 2024-09-07 1489894
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@bf3a03bbb452c2dfb5619b5e4ec1b4ef9fa075cb 🚀
