### 2024-09-09 32a3e4f
tests problems
add badges
tests for select_h function
