### 2024-09-09 af06471
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@32a3e4fd1dc972db3b0ccad46b628cd9bd739e99 🚀
