### 2024-09-09 841eda6
select_h
Trying fo fix the error arising when checking the package or when testing the function select_h.
The issue arises when these processes are run when the package is not installed.
