+++ R/RcppExports.R
@@ -16,7 +16,7 @@
 #' 
 #' @noRd
 computeKernelMatrix <- function(x_mat, y_mat, H) {
-    .Call('_QuadratiK_computeKernelMatrix', PACKAGE = 'QuadratiK', x_mat, y_mat, H)
+    .Call(`_QuadratiK_computeKernelMatrix`, x_mat, y_mat, H)
 }
 
 #' Compute the Poisson kernel matrix between observations in a sample.
@@ -32,7 +32,7 @@ computeKernelMatrix <- function(x_mat, y_mat, H) {
 #' 
 #' @noRd
 computePoissonMatrix <- function(x_mat, rho) {
-    .Call('_QuadratiK_computePoissonMatrix', PACKAGE = 'QuadratiK', x_mat, rho)
+    .Call(`_QuadratiK_computePoissonMatrix`, x_mat, rho)
 }
 
 #' Non-parametric centered kernel
@@ -53,7 +53,7 @@ computePoissonMatrix <- function(x_mat, rho) {
 #' 
 #' @noRd
 NonparamCentering <- function(kmat_zz, n_z) {
-    .Call('_QuadratiK_NonparamCentering', PACKAGE = 'QuadratiK', kmat_zz, n_z)
+    .Call(`_QuadratiK_NonparamCentering`, kmat_zz, n_z)
 }
 
 #' Parametric centered kernel
@@ -75,7 +75,7 @@ NonparamCentering <- function(kmat_zz, n_z) {
 #' 
 #' @noRd
 ParamCentering <- function(kmat_zz, z_mat, H, mu_hat, Sigma_hat) {
-    .Call('_QuadratiK_ParamCentering', PACKAGE = 'QuadratiK', kmat_zz, z_mat, H, mu_hat, Sigma_hat)
+    .Call(`_QuadratiK_ParamCentering`, kmat_zz, z_mat, H, mu_hat, Sigma_hat)
 }
 
 #' Compute kernel-based quadratic distance test for Normality
@@ -87,13 +87,13 @@ ParamCentering <- function(kmat_zz, z_mat, H, mu_hat, Sigma_hat) {
 #' 
 #' @return A scalar value representing the test statistic.
 #'
-#' @useDynLib QuadratiK
+#' @useDynLib QuadratiK, .registration = TRUE
 #' @rdname kbNormTest
 #' @keywords internal
 #' 
 #' @noRd
 kbNormTest <- function(x_mat, h, mu_hat, Sigma_hat) {
-    .Call('_QuadratiK_kbNormTest', PACKAGE = 'QuadratiK', x_mat, h, mu_hat, Sigma_hat)
+    .Call(`_QuadratiK_kbNormTest`, x_mat, h, mu_hat, Sigma_hat)
 }
 
 #' Poisson kernel-based test for Uniformity on the Sphere
@@ -111,7 +111,7 @@ kbNormTest <- function(x_mat, h, mu_hat, Sigma_hat) {
 #' 
 #' @noRd
 statPoissonUnif <- function(x_mat, rho) {
-    .Call('_QuadratiK_statPoissonUnif', PACKAGE = 'QuadratiK', x_mat, rho)
+    .Call(`_QuadratiK_statPoissonUnif`, x_mat, rho)
 }
 
 #'
@@ -127,7 +127,7 @@ statPoissonUnif <- function(x_mat, rho) {
 #' 
 #' @keywords internal
 var_two <- function(Kcen, nsamples) {
-    .Call('_QuadratiK_var_two', PACKAGE = 'QuadratiK', Kcen, nsamples)
+    .Call(`_QuadratiK_var_two`, Kcen, nsamples)
 }
 
[…]
