### 2024-09-09 47c5a43
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@841eda690cf5ac066b4debc9a6368db0695eb793 🚀
