### 2024-09-09 1d340c2
select_h
default number of cores = 2 when checking the package for CRAN
