+++ R/h_selection.R
@@ -337,7 +337,11 @@ select_h <- function(x, y=NULL, alternative=NULL, method="subsampling", b=0.8,
       return(c(STATISTIC[1] < CV))
    }
    
-   if(is.null(n_cores)) {
+   chk <- Sys.getenv("_R_CHECK_LIMIT_CORES_", "")
+   if (nzchar(chk) && chk == "TRUE") {
+      # use 2 cores in CRAN/Travis/AppVeyor
+      num_cores <- 2L
+   } else if(is.null(n_cores)) {
       num_cores <- detectCores()
    } else if (is.numeric(n_cores)) {
       num_cores <- as.numeric(n_cores)
