### 2024-09-09 1b42c22
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@1d340c20235ff163ba6e3a100f3cfec1ec60cab0 🚀
