+++ .github/workflows/pkgcheck.yaml
@@ -13,7 +13,6 @@ on:
     branches:
       - main
       - master
-      - review-rOpenSci
 
 jobs:
   pkgcheck:
