### 2024-09-09 47963d2
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@514d26d6348e651f6a977d2955645c49717c7df0 🚀
