### 2024-09-09 0bc4982
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@1eed1f9c45bb4c36c13e3375c191141346f6421e 🚀
