### 2024-09-09 8a9e784
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@f65d2829eeabca8fd22d549f931d6bc7557db568 🚀
