### 2024-09-09 4e9c2c0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@c65e9b7d8fa7a40712a6ef8c30a9c1d74592f3bb 🚀
