### 2024-09-09 f1d83c7
resolve NOTE
-Add CITATION
-remove vignettes from .Rbuildignore
-resolve NOTE for missing global variable
