+++ R/classes.R
@@ -83,7 +83,7 @@ setClass("kb.test",
 #' @examples
 #' # create a pk.test object
 #' d=3
-#' size=200
+#' size=100
 #' x_sp <- sample_hypersphere(d, n_points=size)
 #' pk.test(x_sp,rho=0.8)
 #'
