### 2024-09-09 3d45acf
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@f1d83c7ba4f6de23580ce0330b6b14a586085f68 🚀
