### 2024-09-10 d41e9a3
testing
- addition of tests for all the functions
- try if the vignettes on 2 and k sample tests can be build on R CMD CHECK on github
