+++ tests/testthat/test-select_h.R
@@ -26,6 +26,23 @@ test_that("Error on invalid method input", {
                 "x must be numeric", fixed=TRUE)
    
    x <- matrix(rnorm(100), ncol = 2)
+   x[1,] <- NA
+   expect_error(select_h(x = x, alternative="skewness"), 
+                "There are missing values in x!", fixed=TRUE)
+   x[1,] <- Inf
+   expect_error(select_h(x = x, alternative="skewness"), 
+                "There are undefined values in x, that is Nan, Inf, -Inf", fixed=TRUE)
+   
+   x <- matrix(rnorm(100), ncol = 2)
+   y <- matrix(rnorm(100), ncol = 2)
+   y[1,] <- NA
+   expect_error(select_h(x = x, y = y, alternative="skewness"), 
+                "There are missing values in y!", fixed=TRUE)
+   y[1,] <- Inf
+   expect_error(select_h(x = x, y = y, alternative="skewness"), 
+                "There are undefined values in y, that is Nan, Inf, -Inf", fixed=TRUE)
+   
+   x <- matrix(rnorm(100), ncol = 2)
    y <- matrix(rnorm(90), ncol = 3)
    expect_error(select_h(x, y, alternative="skewness"),
                 "'x' and 'y' must have the same number of columns.", 
@@ -41,6 +58,11 @@ test_that("Error on invalid method input", {
    "delta_dim must be 1 or a numeric vector of length equal to the 
               number of columns of pooled.", fixed=TRUE)
    
+   x <- matrix(rnorm(100), ncol = 2)
+   y <- matrix(rnorm(100), ncol = 2)
+   expect_error(select_h(x, y, alternative="skewness", h_values = c("a","b","c")), "h_values must be a numeric vector", fixed=TRUE)
+   expect_error(select_h(x, y, alternative="skewness", delta = c("a","b","c")), "delta must be a numeric vector", fixed=TRUE)
+   
 })
 
 
@@ -48,7 +70,7 @@ test_that("Error on invalid method input", {
 test_that("Select h", {
    set.seed(123)
    # normality
-   result <- select_h(x = matrix(rnorm(20),ncol=2), alternative="location")
+   result <- select_h(x = as.data.frame(matrix(rnorm(20),ncol=2)), alternative="location")
    expect_equal(class(result$h_sel), "numeric")
    expect_equal(class(result$power), "data.frame")
 
