### 2024-09-10 a074d37
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@d41e9a3d870edf13913b49f9b3576928ce2574ad 🚀
