### 2024-09-10 2c3895b
Introduction vignette
- fix bug pkbc_validation when true labels are not provided.
