+++ R/QuadratiK-package.R
@@ -52,8 +52,9 @@
 #' 
 #' @note
 #' The `QuadratiK` package is also available in Python on PyPI 
-#' <https://pypi.org/project/QuadratiK/> and also as a Dashboard application 
-#' (ADD LINK).
+#' <https://pypi.org/project/QuadratiK/> and also as a Dashboard application.
+#' Usage instruction for the Dashboard can be found at 
+#' <https://quadratik.readthedocs.io/en/latest/user_guide/dashboard_application_usage.html>.
 #' 
 #' @author Giovanni Saraceno, Marianthi Markatou, 
 #'          Raktim Mukhopadhyay, Mojgan Golzy
