+++ README.md
@@ -31,7 +31,7 @@ library(devtools)
 install_github('giovsaraceno/QuadratiK-package')
 ```
 
-The `QuadratiK` package is also available in Python on PyPI <https://pypi.org/project/QuadratiK/> and also as a Dashboard application (ADD LINK).
+The `QuadratiK` package is also available in Python on PyPI <https://pypi.org/project/QuadratiK/> and also as a Dashboard application. Usage instruction for the Dashboard can be found at <https://quadratik.readthedocs.io/en/latest/user_guide/dashboard_application_usage.html>.
 
 ## Authors
 
