### 2024-09-10 0a29dd8
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@2c3895b380733b027c07e5e6dedacef65d4953c7 🚀
