+++ tests/testthat/test-rpkb.R
@@ -103,23 +103,26 @@ test_that("Random Generation from PKBD works", {
 test_that("Random Generation from PKBD compared to wrapped Cauchy", {
    
    require(circular)
-   
-   # Parameters 
-   n <- 1000  
-   location <- circular(0)  
-   rho <- 0.5  
-   # Generate data from wrapped cauchy
-   wc <- circular::rvonmises(n, mu = location, kappa = -log(rho))
-   # Generate data from pkbd 
-   pkbd <- rpkb(n, mu = c(1, 0), rho = 0.5)$x
-   
-   # Convert Cartesian coordinates to angles for comparison
-   pkbd_angles <- atan2(pkbd[,2], pkbd[,1]) %% (2*pi)
-   
-   # Perform Kolmogorov-smirnov test 
-   # (common for comparing two circular distributions)
-   ks_test <- ks.test(pkbd_angles, wc)
-   expect_lt(ks_test$p.value,0.05)
 
+   # Parameters 
+   n <- 10000
+   location <- circular::circular(pi)  
+   rho <- 0.6
+   set.seed(1234)
+   # Generate data from wrapped cauchy
+   wc <- circular::rwrappedcauchy(n, mu = location, rho = rho)
+   # Generate data from pkbd 
+   pkbd <- rpkb(n, mu = c(-1, 0), rho = rho)$x
+   # Convert Cartesian coordinates to angles for comparison
+   pkbd_angles <- circular::coord2rad(pkbd, control.circular = list(modulo="asis"))
+   
+   # Compute the sample circular quantiles 
+   vect_quantiles <- c(0.05,seq(0.1,0.9,0.05), 0.95)
+   wc_q <- circular::quantile.circular(wc, vect_quantiles)
+   pkbd_q <- circular::quantile.circular(pkbd_angles, vect_quantiles)
+   
+   # Test that the computed quantiles are close
+   tol <- 0.01
+   expect_equal(wc_q, pkbd_q, tolerance = tol)
    
 })
