### 2024-09-12 405fecb
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@9ffcfda68f1ff82f967c6dd86835cc8b4c5f208d 🚀
