### 2024-09-12 9326702
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@ea89203a252abffe2df9eb5b3d9ff206f2c0d580 🚀
