### 2024-09-12 1fe9125
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@c480f2e230cdd16a44a5a62b101d26060ec3791f 🚀
