### 2024-09-12 9d66329
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@bb51762528080f302b0a0bb6caf41d3c987bc865 🚀
