+++ pkgdown.yml
@@ -8,4 +8,4 @@ articles:
   TwoSample_test: TwoSample_test.html
   uniformity: uniformity.html
   wireless_clustering: wireless_clustering.html
-last_built: 2024-09-12T13:36Z
+last_built: 2024-09-12T13:47Z
