### 2024-09-12 528d02a
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@c42c5658b48f868541b2da002d9bfc722c5f3108 🚀
