### 2024-09-12 1ce71be
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@a25c945abca1cce0572e20ce63612e34ca195382 🚀
