### 2024-09-12 99bb123
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@55f9b0263ebd31808bca66f49c18ea9b221594f7 🚀
