### 2024-09-12 922c303
update NEWS
