### 2024-09-12 f828ee9
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@922c3036a4e56a95eb78567e8b6fcf9182bd71eb 🚀
