### 2024-09-12 b3115bc
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@f557d0e893b5ced37b7a76a4e5081e7b7c42fa6c 🚀
