### 2024-09-18 684fb9d
vignettes
- Change the vignette adding some formulas and some explaination of the chosen examples.
- help documentation for the methods of the classes
- update NEWS
