### 2024-09-18 7b16c77
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@8a79252089b2e39eb24a12497c73b542bc192464 🚀
