### 2024-09-18 1f20949
lintr
good practice for coding
