### 2024-09-18 939f888
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@1f20949416e4e0662bd14aa5dc1b06f37abd52c1 🚀
