### 2024-09-18 7bc9fd9
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@64b6bce048a7964190a5a8f8211ff3d213830465 🚀
