### 2024-09-20 d907dd7
testing
reach 100% coverage
- The portions of code not covered are those asking the installation of suggested packages
