+++ tests/testthat/test-kb.test.R
@@ -58,6 +58,28 @@ test_that("Error on invalid centeringType input", {
 # Test 5: Correct handling of vector x input
 test_that("Handle vector x input correctly", {
    
+   # NA in the data
+   datx <- matrix(rnorm(100),ncol=2)
+   daty <- matrix(rnorm(100),ncol=2)
+   daty[1,] <- NA 
+   expect_error(kb.test(x = datx, y = daty, h = 0.8), 
+                'There are missing values in y!', fixed=TRUE)
+   datx[1,] <- NA 
+   expect_error(kb.test(x = datx, h = 0.8), 
+                'There are missing values in x!', fixed=TRUE)
+   
+   # Inf or Nan in the data
+   datx <- matrix(rnorm(100),ncol=2)
+   daty <- matrix(rnorm(100),ncol=2)
+   daty[1,] <- Inf 
+   expect_error(kb.test(x = datx, y = daty, h = 0.8), 
+                'There are undefined values in y, that is Nan, Inf, -Inf', 
+                fixed=TRUE)
+   datx[1,] <- Inf 
+   expect_error(kb.test(x = datx, h = 0.8), 
+                'There are undefined values in x, that is Nan, Inf, -Inf', 
+                fixed=TRUE)
+   
    set.seed(123)
    # x is a vector
    result <- kb.test(x = rnorm(10), h=0.5)
@@ -71,6 +93,11 @@ test_that("Handle vector x input correctly", {
    # x is a matrix
    result <- kb.test(x = matrix(rnorm(20),ncol=2), h=0.5)
    expect_s4_class(result, "kb.test")
+   
+   # test show method
+   output <- capture.output(show(result))
+   expect_true(any(grepl("\t\tU-statistic\tV-statistic", output)))
+   expect_true(any(grepl("H0 is rejected:\t", output)))
    
    # test summary method
    s <- summary(result)
