### 2024-09-20 71e40f1
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@d907dd75989df8c6eb2c1b4360f3bedc32aa3d48 🚀
