### 2024-09-20 b8807ad
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@8868065b3b209359addc7a7272dbcb1a090949c2 🚀
