### 2024-09-20 95257ac
update n_cores
- in the previous commit new badges have been added
