### 2024-09-20 c99e485
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@95257ac7005bfeecc5cb9c5e9ae991884414343a 🚀
