### 2024-09-20 014357a
fix build vignettes
