+++ vignettes/Introduction.R
@@ -19,7 +19,7 @@ kb.test(x, h=0.4)
 
 ## -----------------------------------------------------------------------------
 x <- matrix(rnorm(100,4), ncol = 2)
-# Does x come from a multivariate standard normal distribution?
+# Does x come from the specified multivariate normal distribution?
 kb.test(x, mu_hat = c(4,4), Sigma_hat = diag(2), h = 0.4)
 
 ## -----------------------------------------------------------------------------
