### 2024-09-20 54d74ac
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@014357aeb0a60e53e247ad16b15d7aed0d539971 🚀
