### 2024-09-20 7798dfd
add parallel
add parallel in DESCRIPTION
