+++ DESCRIPTION
@@ -26,6 +26,7 @@ BugReports: https://github.com/giovsaraceno/QuadratiK-package/issues
 Depends: 
     R (>= 3.5.0)
 Imports: 
+    parallel,
     doParallel,
     foreach,
     ggplot2,

[…]
