### 2024-09-20 1127fd4
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@659b99814da1edca93fa38bb8094d7052df0b6b1 🚀
