### 2024-09-20 c2339cd
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@64dfbcd776151d85f7801e2c6ea775ac40f7ce78 🚀
