+++ .github/workflows/test-coverage.yaml
@@ -28,11 +28,7 @@ jobs:
 
       - name: Test coverage
         run: |
-          covr::codecov(
-            quiet = FALSE,
-            clean = FALSE,
-            install_path = file.path(normalizePath(Sys.getenv("RUNNER_TEMP"), winslash = "/"), "package")
-          )
+          covr::codecovtoken = "${{ secrets.CODECOV_TOKEN }}")
         shell: Rscript {0}
 
       - name: Show testthat output
