+++ .github/workflows/test-coverage.yaml
@@ -28,7 +28,7 @@ jobs:
 
       - name: Test coverage
         run: |
-          covr::codecovtoken = "${{ secrets.CODECOV_TOKEN }}")
+          covr::codecovtoken = "${{ secrets.CODECOV_TOKEN }}"
         shell: Rscript {0}
 
       - name: Show testthat output
