+++ .github/workflows/test-coverage.yaml
@@ -23,12 +23,11 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::covr
-          needs: coverage
-
+          extra-packages: covr
+   
       - name: Test coverage
         run: |
-          covr::codecovtoken = "${{ secrets.CODECOV_TOKEN }}"
+          covr::codecov(token = "${{ secrets.CODECOV_TOKEN }}")
         shell: Rscript {0}
 
       - name: Show testthat output
