### 2024-09-23 abe6c39
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@734775ea91e841913738b95830798e01d8581c87 🚀
