### 2024-09-23 d5619a5
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@809edab6075d305c790be49a0c447e76f9bc552d 🚀
