### 2024-09-24 24ea4f8
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@a0170b7b3bc1815233e6bea6ae953e457db2e87a 🚀
