+++ CONTRIBUTING.md
@@ -1,76 +1,65 @@
 # Contributing to QuadratiK package
-_Version: 1.0.0_
 
+*Version: 1.1.2*
 
-First of all, thanks for considering contributing to QuadratiK! 👍 4
+First of all, thanks for considering contributing to `QuadratiK`! 👍
 
-QuadratiK is a comprehensive statistical analysis package providing a comprehensive set of goodness-of-fit tests and a clustering technique for spherical data using kernel-based quadratic distances.
-
-[repo]: https://github.com/giovsaraceno/QuadratiK-package
-[issues]: https://github.com/giovsaraceno/QuadratiK-package/issues
-[new_issue]: https://github.com/giovsaraceno/QuadratiK-package/issues/new
-[website]: https://github.com/giovsaraceno/QuadratiK-package
-[citation]: https://arxiv.org/abs/2402.02290
-[email]: mailto:gsaracen@buffalo.edu
+`QuadratiK` is a comprehensive statistical analysis package providing a comprehensive set of goodness-of-fit tests and a clustering technique for spherical data using kernel-based quadratic distances.
 
 ## Code of conduct
 
 Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
 
-
 ### Share the love ❤️
 
-Think QuadratiK is useful? Let others discover it, by telling them in person, via Twitter, ResearchGate or a blog post.
+Think `QuadratiK` is useful? Let others discover it, by telling them in person, via Twitter, ResearchGate or a blog post.
 
-Using QuadratiK for a paper you are writing? [Cite it][citation].
+Using `QuadratiK` for a paper you are writing? [Cite it](https://arxiv.org/abs/2402.02290).
 
 ### Ask a question
 
-Using QuadratiK and got stuck? Browse the [documentation][website] to see if you can find a solution. Post your questions and requests as an [issue on GitHub][new_issue]. We'll try to do our best to address it, as questions often lead to better documentation or the discovery of bugs.
+Using `QuadratiK` and got stuck? Browse the [documentation](https://github.com/giovsaraceno/QuadratiK-package) to see if you can find a solution. Post your questions and requests as an [issue on GitHub](https://github.com/giovsaraceno/QuadratiK-package/issues/new). We'll try to do our best to address it, as questions often lead to better documentation or the discovery of bugs.
 
-Want to ask a question in private? Contact the package maintainer by [email][mailto:email].
+Want to ask a question in private? Contact the package maintainer by [gsaracen\@buffalo.edu](mailto:gsaracen@buffalo.edu).
 
 ### Propose an idea 💡
 
-Have an idea for a new QuadratiK feature? 
-Take a look at the [documentation][website] and [issue list][issues] to see if it isn't included or suggested yet. If not, suggest your idea as an [issue on GitHub][new_issue]. Please
+Have an idea for a new `QuadratiK` feature? Take a look at the [documentation](https://github.com/giovsaraceno/QuadratiK-package) and [issue list](https://github.com/giovsaraceno/QuadratiK-package/issues) to see if it isn't included or suggested yet. If not, suggest your idea as an [issue on GitHub](https://github.com/giovsaraceno/QuadratiK-package/issues/new). Please
 
-* Explain in detail how it would work.
-* Keep the scope as narrow as possible.
-* Functions are described as comments near their code and translated to documentation using [`roxygen2`](https://klutometis.github.io/roxygen/). Write documentation using roxygen2 package.
-* Test your changes with goodpractice before submitting.
+-   Explain in detail how it would work.
+-   Keep the scope as narrow as possible.
[…]
