### 2024-09-24 faec157
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@1b7e381f7c2da0b90c5c4c35a00797a6b8ef5e45 🚀
