### 2024-09-26 8f2413b
standardize references
also fix badges links
