+++ R/clustering_functions.R
@@ -99,7 +99,7 @@
 #'   \code{input}: List of input information.
 #'
 #' @references
-#' Golzy, M., Markatou, M. (2020) Poisson Kernel-Based Clustering on the 
+#' Golzy, M. and Markatou, M. (2020) Poisson Kernel-Based Clustering on the 
 #' Sphere: Convergence Properties, Identifiability, and a Method of Sampling, 
 #' Journal of Computational and Graphical Statistics, 29:4, 758-770, 
 #' DOI: 10.1080/10618600.2020.1740713.
@@ -1045,7 +1045,7 @@ setMethod("predict", signature(object="pkbc"),
 #'          \linkS4class{pkbc} for the class object definition.
 #'
 #' @references
-#' Kapp, A.V., Tibshirani, R. (2007) "Are clusters found in one dataset present 
+#' Kapp, A.V. and Tibshirani, R. (2007) "Are clusters found in one dataset present 
 #' in another dataset?", Biostatistics, 8(1), 9–31, 
 #' https://doi.org/10.1093/biostatistics/kxj029
 #' 
