+++ vignettes/uniformity.Rmd
@@ -121,6 +121,6 @@ In this example, the Poisson kernel-based test statistics reject the null hypoth
 
 ## References
 
-Ding, Y., Markatou, M., Saraceno, G. (2023). "Poisson Kernel-Based Tests for
+Ding, Y., Markatou, M. and Saraceno, G. (2023). "Poisson Kernel-Based Tests for
 Uniformity on the d-Dimensional Sphere." *Statistica Sinica*. 
 [doi:10.5705/ss.202022.0347](https://doi.org/10.5705/ss.202022.0347).

[…]
