### 2024-09-26 3f70404
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@8f2413baea998d0735eb3be0682c644f8ded6084 🚀
