### 2024-09-26 1c56abe
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@14c8428cfff71a8c5d5bf1f4493621ff66b85ff4 🚀
