### 2024-09-26 b728dda
spell_check
Add link to introductory vignette in README and in package help documenation
