### 2024-09-26 eb79244
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@b728dda2f235c042cc1069d90ee354ec185fb1e6 🚀
