### 2024-09-26 857e568
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@fe46e641e27b6aa3ec761ac03f6af2f9a58a9b8c 🚀
