### 2024-09-26 29cc5da
details
remove rho=NULL in pk.test
add best practice
add cases in which use pk.test
