### 2024-09-26 6ea8fb0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@29cc5da6c64513fa4c4de778ddfe1608e521862f 🚀
