### 2024-10-27 34c9a16
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@238adcdae30a6c45d4b1c3c7e758fe7dff4e7752 🚀
