+++ man/QuadratiK-package.Rd
@@ -91,7 +91,7 @@ https://doi.org/10.48550/arXiv.2407.16374
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://cran.r-project.org/package=QuadratiK}
+  \item \url{https://CRAN.R-project.org/package=QuadratiK}
   \item \url{https://github.com/giovsaraceno/QuadratiK-package}
   \item \url{https://giovsaraceno.github.io/QuadratiK-package/}
   \item Report bugs at \url{https://github.com/giovsaraceno/QuadratiK-package/issues}
