### 2024-10-29 275bc32
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@32b794e7d0a3e5faf4073f89e7c6c85834e6b42b 🚀
