### 2024-11-14 #632 @kasselhingee
Hi @giovsaraceno the package is much improved with much more information in the documentation. I noticed a few small things to change and some queries, but I do not need to see it again. Once the small changes have been made I recommend this package for Gold badge.

The below is a report I generated with the help of `pkgreviewr::pkgreview_create()`.

# `QuadratiK` - package review

## **Reviewer:** [\@kasselhingee](https://github.com/kasselhingee)

### Review Submitted:
****

***

This report contains documents associated with the review of **rOpenSci** submitted package:

## Package info

**Description:**

It includes test for multivariate normality, test for uniformity on the d-dimensional
    Sphere, non-parametric two- and k-sample tests, random generation of points from the Poisson
    kernel-based density and clustering algorithm for spherical data. For more information see
    Saraceno G., Markatou M., Mukhopadhyay R. and Golzy M. (2024)
    <doi:10.48550/arXiv.2402.02290>
    Markatou, M. and Saraceno, G. (2024) <doi:10.48550/arXiv.2407.16374>,
    Ding, Y., Markatou, M. and Saraceno, G. (2023) <doi:10.5705/ss.202022.0347>,
    and Golzy, M. and Markatou, M. (2020) <doi:10.1080/10618600.2020.1740713>.

**Author:** Giovanni Saraceno <gsaracen@buffalo.edu> [aut, cre] (ORCID 000-0002-1753-2367), Marianthi Markatou [aut], Raktim Mukhopadhyay [aut], Mojgan Golzy [aut]

**repo url:** <https://github.com/giovsaraceno/QuadratiK-package>

**website url:** <https://giovsaraceno.github.io/QuadratiK-package/>

## Review info

#### See [reviewer guidelines](https://ropensci.github.io/dev_guide/reviewerguide.html) for further information on the rOpenSci review process.

**key review checks:**

- Does the code comply with **general principles in the [Mozilla reviewing guide](https://mozillascience.github.io/codeReview/review.html)**?
- Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
- Are there **improvements** that could be made to the **code style?**
- Is there **code duplication** in the package that should be reduced?
- Are there **user interface improvements** that could be made?
- Are there **performance improvements** that could be made?
- Is the [**documentation**](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) **clear and sufficient**?

Please be respectful and kind to the authors in your reviews. The rOpenSci [code of conduct](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) is mandatory for everyone involved in our review process.

***

## Outcome
The package is much improved with much more information in the documentation. I noticed a few small things to change and some queries, but I do not need to see it again. Once the small changes have been made I recommend this package for Gold badge.

## High Level Check List

Estimated hours spent reviewing: 6 hours this time - there are so many things to check!

### session info

``` r
#> R version 4.4.1 (2024-06-14)
#> Platform: x86_64-pc-linux-gnu
#> Running under: Ubuntu 24.04.1 LTS
#>
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.12.0
#> LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.12.0
#>
#> locale:
#>  [1] LC_CTYPE=en_AU.UTF-8       LC_NUMERIC=C
#>  [3] LC_TIME=en_AU.UTF-8        LC_COLLATE=en_AU.UTF-8
#>  [5] LC_MONETARY=en_AU.UTF-8    LC_MESSAGES=en_AU.UTF-8
#>  [7] LC_PAPER=en_AU.UTF-8       LC_NAME=C
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C
#> [11] LC_MEASUREMENT=en_AU.UTF-8 LC_IDENTIFICATION=C
#>
#> time zone: Australia/Sydney
#> tzcode source: system (glibc)
#>
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base
#>
#> other attached packages:
#> [1] devtools_2.4.5     usethis_3.0.0.9000 magrittr_2.0.3
#>
#> loaded via a namespace (and not attached):
#>  [1] miniUI_0.1.1.1    jsonlite_1.8.9    compiler_4.4.1    promises_1.3.0
#>  [5] Rcpp_1.0.13-1     later_1.3.2       jquerylib_0.1.4   yaml_2.3.10
#>  [9] fastmap_1.2.0     mime_0.12         R6_2.5.1          knitr_1.49
#> [13] htmlwidgets_1.6.4 profvis_0.4.0     shiny_1.9.1       bslib_0.8.0
#> [17] rlang_1.1.4       cachem_1.1.0      httpuv_1.6.15     xfun_0.49
#> [21] fs_1.6.5          sass_0.4.9        pkgload_1.4.0     memoise_2.0.1
#> [25] cli_3.6.3         digest_0.6.37     rstudioapi_0.17.1 xtable_1.8-4
#> [29] remotes_2.5.0     lifecycle_1.0.4   vctrs_0.6.5       evaluate_1.0.1
#> [33] glue_1.8.0        urlchecker_1.0.1  sessioninfo_1.2.2 pkgbuild_1.4.5
#> [37] rmarkdown_2.29    purrr_1.0.2       tools_4.4.1       ellipsis_0.3.2
#> [41] htmltools_0.5.8.1
```

## Test installation

### test local `QuadratiK` install:

``` r
install(pkg_dir, dependencies = T, build_vignettes = T)
```

``` r
remove.packages("QuadratiK")
#> Removing package from '/home/kassel/R/x86_64-pc-linux-gnu-library/4.4'
#> (as 'lib' is unspecified)
```
#### **comments:**

Local install was seamless.

***

### test install of `QuadratiK` from GitHub with:

``` r
devtools::install_github("giovsaraceno/QuadratiK-package", dependencies = T, build_vignettes = T)
#> Downloading GitHub repo giovsaraceno/QuadratiK-package@HEAD
#>
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#>
   checking for file ‘/tmp/Rtmpix2IQB/remotesa0043e3633a2/giovsaraceno-QuadratiK-package-32b794e/DESCRIPTION’ ...

✔  checking for file ‘/tmp/Rtmpix2IQB/remotesa0043e3633a2/giovsaraceno-QuadratiK-package-32b794e/DESCRIPTION’
#>

─  preparing ‘QuadratiK’:
#>    checking DESCRIPTION meta-information ...

✔  checking DESCRIPTION meta-information
#> ─  cleaning src
#>

#>

   creating vignettes ...

✔  creating vignettes (1m 7.6s)
#>

─  cleaning src
#>

─  checking for LF line-endings in source and make files and shell scripts (685ms)
#>

#>

─  building ‘QuadratiK_1.1.2.tar.gz’
#>

#>
#> Installing package into '/home/kassel/R/x86_64-pc-linux-gnu-library/4.4'
#> (as 'lib' is unspecified)
```

#### **comments:**

Installing from GitHub works.

***

## Check package integrity

### run checks on `QuadratiK` source:

``` r
devtools::check(pkg_dir)
```
#### **comments:**

Checks pass.

***

### run tests on `QuadratiK` source:

#### **comments:**

Tests pass.

***

### check `QuadratiK` for goodpractice:

``` r
goodpractice::gp(pkg_dir)
#> ℹ Preparing: covr
#> ℹ Preparing: cyclocomp
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#>
   checking for file ‘/tmp/Rtmpp2nRfX/remotesb5005dc86caa/QuadratiK/DESCRIPTION’ ...

✔  checking for file ‘/tmp/Rtmpp2nRfX/remotesb5005dc86caa/QuadratiK/DESCRIPTION’
#>

─  preparing ‘QuadratiK’:
#>

   checking DESCRIPTION meta-information ...

✔  checking DESCRIPTION meta-information
#>

─  cleaning src
#>

   checking vignette meta-information ...

✔  checking vignette meta-information
#>

─  checking for LF line-endings in source and make files and shell scripts (880ms)
#>

#>

─  building ‘QuadratiK_1.1.2.tar.gz’
#>

#>
#> ℹ Preparing: description
#> ℹ Preparing: lintr
#> ℹ Preparing: namespace
#> ℹ Preparing: rcmdcheck
#> ── GP QuadratiK ────────────────────────────────────────────────────────────────
#>
#> It is good practice to
#>
#>   ✖ avoid long code lines, it is bad for readability. Also, many people prefer
#>     editor windows that are about 80 characters wide. Try make your lines
#>     shorter than 80 characters
#>
#>     'R/clustering_functions.R:1048:81'
#>     'R/QuadratiK-package.R:62:81'
#>     'tests/testthat/test-pkbc.R:199:81'
#>     'tests/testthat/test-select_h.R:93:81'
#>
#>   ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...)
#>     expressions. They are error prone and result 1:0 if the expression on the
#>     right hand side is zero. Use seq_len() or seq_along() instead.
#>
#>     'R/h_selection.R:366:28'
#>
#>   ✖ not import packages as a whole, as this can cause name clashes between the
#>     imported packages, especially over time as packages change. Instead, import
#>     only the specific functions you need.
#>   ✖ fix this R CMD check NOTE: installed size is 17.3Mb sub-directories of 1Mb
#>     or more: doc 1.0Mb libs 15.3Mb
#> ────────────────────────────────────────────────────────────────────────────────
```
#### **comments:**

I agree with the above suggestions, but not sure if they are part of rOpenSci standards. Particularly the suggestion about reducing size I would ignore (though I am surprised `libs` is so large).

***

## Check package metadata files

### inspect

- #### [README](https://github.com/giovsaraceno/QuadratiK-package)
- #### [DESCRIPTION](https://github.com/giovsaraceno/QuadratiK-package/blob/master/DESCRIPTION)
- #### [NAMESPACE](https://github.com/giovsaraceno/QuadratiK-package/blob/master/NAMESPACE)

+ is `d` the dimension of the ambient Euclidean space or the dimension of the sphere? Seems like you're using it for both which I found hard to untangle.
+ Is the Poisson kernal's $x$ on the sphere too? ($d$-variate makes it sound like Euclidean space)

+ `x` matrix of data points: please specify whether each data point a row in the matrix (like in tidyverse) or a column?

+ Namespace has a duplication. I'm not sure what R would actually be doing in this case.
```
useDynLib(QuadratiK)
useDynLib(QuadratiK, .registration = TRUE)
```

+ A few packages are fully imported according to the NAMESPACE. This makes sense when you're using alot of the functionality of that package, or a lot of functions from that package. I'd guess that most can be converted to `importFrom()` statements.
```
import(Rcpp)
import(RcppEigen)
import(foreach)
import(ggplot2)
import(rlecuyer)
```

### spell check

``` r
devtools::spell_check(pkg_dir)
#>   WORD               FOUND IN
#> Aeberhard          wine.Rd:27,51
#> Ajne               uniformity.Rmd:107,120
#> al                 dpkb.Rd:40,42
#>                    kb.test.Rd:151,162
#>                    plot.pkbc.Rd:36,59
#>                    wireless.Rd:49
#> ARI                pkbc_validation.Rd:31,42,44,66,69
#>                    wireless_clustering.Rmd:42,51
#> arxiv              select_h.Rd:154
#> arXiv              compute_CV.Rd:56
#>                    kb.test.Rd:288
#>                    QuadratiK-package.Rd:75,89
#>                    select_h.Rd:149
#>                    description:5,6
#>                    README.md:5,67,67,88
#>                    Introduction.Rmd:55,236
#> ASW                pkbc_validation.Rd:70
#>                    wireless_clustering.Rmd:42
#> Bhatt              wireless.Rd:21,46
#> Bingham            uniformity.Rmd:107,120
#> biostatistics      pkbc_validation.Rd:95
#> Biostatistics      pkbc_validation.Rd:94
#> cdot               uniformity.Rmd:34
#> cen                uniformity.Rmd:23,34
#> CMD                NEWS.md:13,73
#>                    README.md:5
#> codecov            README.md:5
#> codemeta           NEWS.md:81
#> Coomans            wine.Rd:51
#> cr                 TwoSample_test.Rmd:59
#> dBm                wireless.Rd:35
#>                    wireless_clustering.Rmd:27
#> De                 wine.Rd:51
#> det                kSample_test.Rmd:40
#>                    TwoSample_test.Rmd:45
#> DOF                DOF_norm.Rd:5,15,19
#>                    DOF.Rd:5,15,19
#> doi                breast_cancer.Rd:14
#>                    compute_CV.Rd:56
#>                    kb.test.Rd:288
#>                    pk.test.Rd:119
#>                    pkbc_validation.Rd:95
#>                    plot.pkbc.Rd:61
#>                    poisson_CV.Rd:38
#>                    QuadratiK-package.Rd:79,89
#>                    select_h.Rd:149
#>                    wine.Rd:29
#>                    wireless.Rd:23,51
#>                    description:5,6,7,8
#>                    README.md:84
#>                    Introduction.Rmd:238
#>                    uniformity.Rmd:126
#> DOI                dpkb.Rd:118
#>                    kb.test.Rd:293
#>                    pkbc.Rd:154
#>                    QuadratiK-package.Rd:84
#>                    README.md:86
#>                    generate_rpkb.Rmd:126,130
#>                    Introduction.Rmd:240
#>                    wireless_clustering.Rmd:93
#> DW                 breast_cancer.Rd:14
#> EJS                generate_rpkb.Rmd:130
#> et                 dpkb.Rd:40,42
#>                    kb.test.Rd:151,162
#>                    plot.pkbc.Rd:36,59
#>                    wireless.Rd:49
#> Fi                 wireless.Rd:30,31,35
#>                    wireless_clustering.Rmd:27,29
#> Fk                 var_k.Rd:21
#> FNA                breast_cancer.Rd:22
#> Forina             wine.Rd:27
#> frac               README.md:16
#>                    generate_rpkb.Rmd:29
#>                    kSample_test.Rmd:32,32,40,47,68
#>                    TwoSample_test.Rmd:36,36,36,40,40,45,52
#>                    uniformity.Rmd:23,34
#> ge                 README.md:11
#>                    kSample_test.Rmd:128
#> geospatial         Introduction.Rmd:14
#> ggplot             compare_qq.Rd:5,20
#> gitignore          NEWS.md:84
#> Hornik             dpkb.Rd:80,120
#>                    generate_rpkb.Rmd:48,128
#> https              breast_cancer.Rd:14
#>                    compute_CV.Rd:56
#>                    kb.test.Rd:288
#>                    pkbc_validation.Rd:95
#>                    plot.pkbc.Rd:61
#>                    QuadratiK-package.Rd:89
#>                    select_h.Rd:149,154
#>                    wine.Rd:29
#>                    wireless.Rd:23,51
#> hypersphere        pkbc.Rd:85
#>                    sample_hypersphere.Rd:5,18
#> IGP                pkbc_validation.Rd:35,38,69
#>                    wireless_clustering.Rmd:42
#> ij                 kSample_test.Rmd:32,57
#> InGroup            wireless_clustering.Rmd:42
#> initializers       pkbc.Rd:104
#> interpretability   QuadratiK-package.Rd:55
#>                    README.md:20
#>                    Introduction.Rmd:193
#> json               NEWS.md:81
#> Kaleida            QuadratiK-package.Rd:63
#>                    README.md:92
#> Kapp               pkbc_validation.Rd:30,40,93
#> KBQD               compute_CV.Rd:5
#>                    cv_ksample.Rd:5,44
#>                    kb.test.Rd:7,146
#>                    normal_CV.Rd:5,30,35
#>                    pk.test.Rd:63,72
#>                    poisson_CV.Rd:5
#>                    select_h.Rd:80,86
#>                    TwoSample_test.Rmd:30,59
#>                    uniformity.Rmd:23,34
#> kxj                pkbc_validation.Rd:95
#> ldots              README.md:13
#>                    Introduction.Rmd:122
#>                    kSample_test.Rmd:17,20,47
#>                    TwoSample_test.Rmd:18,59
#>                    uniformity.Rmd:17
#> le                 kSample_test.Rmd:20
#> Leydold            dpkb.Rd:80,120
#>                    generate_rpkb.Rmd:48,128
#> Lifecycle          README.md:5
#> Locantore          plot.pkbc.Rd:36,59
#> logLikelihood      summary.pkbc.Rd:14
#> Mangasarian        breast_cancer.Rd:11,32
#> Marron             plot.pkbc.Rd:59
#> mathbb             README.md:16
#>                    generate_rpkb.Rmd:23,34
#>                    Introduction.Rmd:142
#>                    kSample_test.Rmd:47
#>                    TwoSample_test.Rmd:52
#> mathbf             README.md:16
#>                    generate_rpkb.Rmd:23,29
#>                    kSample_test.Rmd:17,18,30,32,40,47,57
#>                    TwoSample_test.Rmd:35,36,36,40,40,45,52
#>                    uniformity.Rmd:23,25,34,36,48
#> mathcal            README.md:16
#>                    generate_rpkb.Rmd:23
#>                    Introduction.Rmd:93,142
#>                    uniformity.Rmd:17,48
#> mathrm             kSample_test.Rmd:57
#>                    TwoSample_test.Rmd:35
#> mbox               kSample_test.Rmd:32,36,51
#> md                 NEWS.md:6,60,64,77,82
#> Memb               predict.pkbc.Rd:23
#> milliwatts         wireless.Rd:35
#>                    wireless_clustering.Rmd:27
#> Mises              dpkb.Rd:37,79
#>                    generate_rpkb.Rmd:48
#>                    uniformity.Rmd:84
#> Narayanan          wireless.Rd:46
#> Nonparam           kb.test.Rd:78
#>                    NEWS.md:87
#> nonumber           TwoSample_test.Rmd:41
#> Param              kb.test.Rd:78
#> parametrically     kb.test.Rd:190
#> PCALocantore       plot.pkbc.Rd:21
#> Perumal            wireless.Rd:46
#> pkbc               plot.pkbc.Rd:30
#>                    summary.pkbc.Rd:20
#> PKBD               dpkb.Rd:6,52,74,79,81,120
#>                    QuadratiK-package.Rd:31,33
#>                    summary.pkbc.Rd:6
#>                    NEWS.md:16
#>                    README.md:16,16
#>                    generate_rpkb.Rmd:48,128
#>                    Introduction.Rmd:151,153,230
#> PKBDs              dpkb.Rd:71,72
#>                    generate_rpkb.Rmd:34
#> pkgcheck           NEWS.md:73
#> pkgdown            NEWS.md:65
#> poisson            poisson_CV.Rd:26
#> preprint           QuadratiK-package.Rd:75
#>                    README.md:67
#>                    Introduction.Rmd:55
#> Probs              predict.pkbc.Rd:24
#> PyPI               QuadratiK-package.Rd:67
#>                    README.md:39
#>                    Introduction.Rmd:32
#> qq                 summary.kb.test.Rd:20
#>                    summary.pk.test.Rd:20
#>                    TwoSample_test.Rmd:94
#>                    uniformity.Rmd:73,78
#> qquad              kSample_test.Rmd:32,36
#> Rbuildignore       NEWS.md:81,82
#> README             NEWS.md:8,77
#> rejacg             dpkb.Rd:46,48,97
#> Rohra              wireless.Rd:46
#> rOpenSci           NEWS.md:13
#>                    README.md:5
#> Rousseeuw          pkbc_validation.Rd:31,52,97
#> roxygen            NEWS.md:27
#> Sablica            dpkb.Rd:40,42,80,120
#>                    generate_rpkb.Rmd:48,128
#> Satterthwaite      kb.test.Rd:160
#>                    pk.test.Rd:87
#> Sinica             pk.test.Rd:118
#>                    poisson_CV.Rd:38
#>                    QuadratiK-package.Rd:79
#>                    README.md:84
#>                    Introduction.Rmd:238
#>                    uniformity.Rmd:125
#> SPIE               breast_cancer.Rd:34
#> Springer           wireless.Rd:51
#> srr                NEWS.md:61
#> ss                 pk.test.Rd:119
#>                    poisson_CV.Rd:38
#>                    QuadratiK-package.Rd:79
#>                    description:7
#>                    README.md:84
#>                    Introduction.Rmd:238
#>                    uniformity.Rmd:126
#> Statistica         pk.test.Rd:118
#>                    poisson_CV.Rd:38
#>                    QuadratiK-package.Rd:79
#>                    README.md:84
#>                    Introduction.Rmd:238
#>                    uniformity.Rmd:125
#> testthat           NEWS.md:15
#> th                 kb.test.Rd:168
#>                    pk.test.Rd:97
#>                    predict.pkbc.Rd:25
#> Thakur             wireless.Rd:46
#> Tibshirani         pkbc_validation.Rd:30,40,93
#> UCI                breast_cancer.Rd:13
#>                    wine.Rd:28
#>                    wireless.Rd:22
#>                    wireless_clustering.Rmd:20
#> Un                 kb.test-class.Rd:30
#>                    kb.test.Rd:99,100
#>                    pk.test-class.Rd:20,23
#>                    pk.test.Rd:20,35,37,46
#> unbiasedness       kb.test.Rd:259
#>                    pk.test.Rd:106
#> Vel                wine.Rd:51
#> Vn                 kb.test-class.Rd:26,32
#>                    kb.test.Rd:103,104
#>                    pk.test-class.Rd:27,31
#>                    pk.test.Rd:38,39,42
#> von                dpkb.Rd:37,79
#>                    generate_rpkb.Rmd:48
#>                    uniformity.Rmd:84
#> wcss               plot.pkbc.Rd:41,47
#>                    summary.pkbc.Rd:15
#> Wi                 wireless.Rd:30,31,35
[…]
