### 2024-11-15 e94750c
ropensci review 14-11-24
- the function rpkb() returns directly the generated data set and not a list
- packages ggplot2 and foreach not called as full packages but using importFrom
 - nClust argument of pkbc function has no default value
- usage of standardGeneric (see https://adv-r.hadley.nz/s4.html#s4-generics)
- goodpractice sugestions
- Uniform notation: d is dimension in Eucldean space with spherical data lying on the (d-1)-dimensional sphere
- Specify dimension of input matrix x in clustering functions.
- Improve explaination of Average Silhouette Width
- Update NEWS.md file
- Add Hinge
[…]
