+++ tests/testthat/test-validation.R
@@ -1,4 +1,4 @@
-library(testthat)
+
 #------------------------------------------------------
 ## Clustering on the Sphere
 ## 
