### 2024-11-15 0be8fa2
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@e94750ce149e727618db968f4ff42bf5443c2958 🚀
