### 2024-11-21 e496b42
review rOpenSci 2
- motivation L2 normalization wireless data
- specify the rho in pk.test cannot be equal 1
