### 2024-11-21 1562ab0
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@e496b42bc6b45cc96de95859995d3c42b8e7685d 🚀
