### 2024-11-21 b1c055d
Update pkgcheck.yaml
