### 2024-11-21 b61a45e
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@b1c055d74333a34a2ac84202630847d23a4e5329 🚀
