### 2024-11-22 a17380f
Update pkgcheck.yaml
