### 2024-11-22 5b5a02b
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@a17380f5dcf36c4942af2a068951269ac10a1f8a 🚀
