### 2024-11-22 4960eff
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@7f19efb163f774494ed6592941c44ba817c1309e 🚀
