+++ R/utility.R
@@ -10,9 +10,9 @@
 #' @return a list containing the DOF and the coefficient c of the asymptotic
 #' distribution
 #' 
-#' @srrstats {G1.4a} roxigen2 is used
+#' @srrstats {G1.4a} roxygen2 is used
 #' @keywords internal
-DOF <- function(d, rho){
+DOF <- function(d, rho) {
    num_c <- (1+rho^2)/((1-rho^2)^(d-1)) -1
    den_c <- (1+rho)/((1-rho)^(d-1)) -1
    DOF <- (den_c^2)/num_c
@@ -33,7 +33,7 @@ DOF <- function(d, rho){
 #' @return a list containing the DOF and the coefficient c of the asymptotic
 #' distribution
 #' 
-#' @srrstats {G1.4a} roxigen2 is used
+#' @srrstats {G1.4a} roxygen2 is used
 #' @keywords internal
 DOF_norm <- function(Sigma_h, V){
    
@@ -61,7 +61,7 @@ DOF_norm <- function(Sigma_h, V){
 #'
 #' @return the value of computed variance
 #' 
-#' @srrstats {G1.4a} roxigen2 is used
+#' @srrstats {G1.4a} roxygen2 is used
 #' @keywords internal
 var_norm <- function(Sigma_h, V, n){
    
@@ -88,7 +88,7 @@ var_norm <- function(Sigma_h, V, n){
 #' @examples
 #' x_sp <- sample_hypersphere(3,100)
 #'
-#' @srrstats {G1.4} roxigen2 is used
+#' @srrstats {G1.4} roxygen2 is used
 #' 
 #' @export
 sample_hypersphere <- function(d, n_points=1) {
@@ -116,7 +116,7 @@ sample_hypersphere <- function(d, n_points=1) {
 #'
 #' @useDynLib QuadratiK
 #' 
-#' @srrstats {G1.4a} roxigen2 is used 
+#' @srrstats {G1.4a} roxygen2 is used 
 #' @keywords internal
 generate_SN<-function(d, size_x, size_y, mu_x, mu_y, 
                       sigma_x, sigma_y, skewness_y){
@@ -145,7 +145,7 @@ generate_SN<-function(d, size_x, size_y, mu_x, mu_y,
 #' @importFrom ggplot2 ggplot geom_abline geom_line ggtitle theme_minimal
 #' @importFrom ggplot2 theme scale_color_brewer
 #' 
-#' @srrstats {G1.4a} roxigen2 is used
+#' @srrstats {G1.4a} roxygen2 is used
 #' 
 #' @keywords internal
 compare_qq <- function(sample1, sample2, main_title) {
@@ -188,7 +188,7 @@ compare_qq <- function(sample1, sample2, main_title) {
 #'
 #' @return Computed statistics with a plot
 #' 
-#' @srrstats {G1.4a} roxigen2 is used
+#' @srrstats {G1.4a} roxygen2 is used
 #' 
 #' @keywords internal
 compute_stats <- function(var1, var2, var_name,eps=3) {
