+++ R/wine.R
@@ -43,7 +43,7 @@
 #' Pattern Recognition, 27(8), 1065-1077.
 #' 
 #' @srrstats {G1.0} Reference section reports the related literature
-#' @srrstats {G1.4} roxigen2 is used
+#' @srrstats {G1.4} roxygen2 is used
 #' @srrstats {G5.1} the data set is made available
 #' 
 "wine"
