### 2025-01-29 #632 @emitanaka
---

Thank you for your patience @giovsaraceno. This package is very well documented and is a great contribution to goodness-of-fit testing and clustering for spherical data.

I've made a pull request to ensure that the styling and spelling errors are mostly corrected.

I note that `test-kb.test.R` and `test-select_h.R` hangs when I use `devtools::test()` or `devtools::test_active_file()`
but runs fine when I run the test individually in local environment. I'm not sure why and am presuming it is okay since it runs locally.

I too like to recommend for the gold badge, however, just a few minor points if you could address (shown below).

- In `test-kb.test.R` -- test 7 (testing main functionality of k-sample test)  and test 8 (testing selection of h) -- I see that the tests are looking at the return classes but  there is no test for the actual values (test-statistic, critical value calculation, h, etc).  I think you should add tests for these values as well as per G5.4 *Correctness tests*?
- The function `pkbc` has assertions for the input `nClust`, however, no assertions of _single numeric value_ appear to be made for `maxIter` and `numInit`.
- I don't think below is done?
> Standards on line#154 of file R/clustering_functions.R:
> UL2.3 Unsupervised Learning Software should implement pre-processing routines to identify whether aspects of input data are perfectly collinear.*

- Below, I don't think there are labels for cluster groups that can be supplied?

> Standards on line#574 of file R/clustering_functions.R: - stats_clusters()
> - UL3.2 Unsupervised Learning Software for which input data does not generally include labels (such as array-like data with no row names) should provide an additional parameter to enable cases to be labelled.*
