### 2025-01-29 613e706
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@71b851f9407f9d64fc711c0c712f32b126db72b3 🚀
