### 2025-01-29 d3a3823
review2_part1
- add reviewer in DESCRIPTION
- add assertions about numInit and maxIter in pkbc function are single numeric values
