### 2025-01-29 aa02595
review2_part2
- add assertions about input patameter maxIter and numInit
- package check
