+++ vignettes/TwoSample_test.R
@@ -18,12 +18,12 @@ n <- 100
 d <- 4
 skewness_y <- 0.5
 set.seed(2468)
-x_2 <- rmvnorm(n, mean = rep(0,d))
-y_2 <- rmsn(n=n, xi=0, Omega = diag(d), alpha=rep(skewness_y,d))
+x_2 <- rmvnorm(n, mean = rep(0, d))
+y_2 <- rmsn(n = n, xi = 0, Omega = diag(d), alpha = rep(skewness_y, d))
 
 ## -----------------------------------------------------------------------------
 set.seed(2468)
-two_test <- kb.test(x=x_2, y=y_2)
+two_test <- kb.test(x = x_2, y = y_2)
 two_test
 
 ## ----fig.width=6, fig.height=4------------------------------------------------
@@ -41,12 +41,12 @@ summary_two$summary_tables
 
 ## ----eval=FALSE---------------------------------------------------------------
 #  set.seed(2468)
-#  two_test_h <- select_h(x=x_2, y=y_2, alternative = "skewness")
+#  two_test_h <- select_h(x = x_2, y = y_2, alternative = "skewness")
 
 ## -----------------------------------------------------------------------------
-x_pool <- rbind(x_2,y_2)
-y_memb <- rep(c(1,2),each=n)
+x_pool <- rbind(x_2, y_2)
+y_memb <- rep(c(1, 2), each = n)
 h <- two_test@h$h_sel
 set.seed(2468)
-kb.test(x=x_pool, y=y_memb, h=h)
+kb.test(x = x_pool, y = y_memb, h = h)
 
