+++ vignettes/wireless_clustering.R
@@ -14,7 +14,7 @@ wire <- wireless[,-8]
 labels <- wireless[,8]
 wire_norm <- wire/sqrt(rowSums(wire^2))
 set.seed(2468)
-res_pk <- pkbc(as.matrix(wire_norm),3:5)
+res_pk <- pkbc(as.matrix(wire_norm), 3:5)
 
 ## -----------------------------------------------------------------------------
 set.seed(2468)
@@ -32,7 +32,7 @@ plot(res_pk, k = 4)
 plot(res_pk, k=4, true_label = labels)
 
 ## -----------------------------------------------------------------------------
-summary_clust <- stats_clusters(res_pk,4)
+summary_clust <- stats_clusters(res_pk, 4)
 
 ## -----------------------------------------------------------------------------
 summary_clust
