### 2025-01-29 3b84d41
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@aa02595182f8511b1cb6099a1effb3db40954b50 🚀
