### 2025-01-29 3eb00c9
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@5294b89702814a34f1c8f47fb90c6e611542c29e 🚀
