### 2025-01-29 c549a00
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@3f113d4a7b924dc50ab3acd21ce80cb83543fd5e 🚀
