### 2025-01-31 #632 @giovsaraceno
Dear @emitanaka,

Thank you for your thorough review, with the suggestions and changes provided for improving the package, and for recommending the package for the gold badge.
We are grateful for your contributions as a reviewer and agree to acknowledge you as a package reviewer in the DESCRIPTION file.
Below, we address the final minor points raised in the review for clarity and completeness.
On behalf of all the authors, I would like to thank you and the rOpenSci team again for your time and effort in reviewing this package. Please, let me know what is the next step of the review process.

- I have successfully merged your pull request addressing the styling and spelling errors.

-  I have added assertions in the `pkbc` function to ensure that `maxIter` and `numInit` are single numeric values.

- The standard UL2.3 has been removed from the mentioned line.

- Standard UL3.2, (about the option of providing the true labels as a separate input, if available) has been removed from the function `stats_clusters()`, since the computed statistics do not depend on true labels. The standard tag has been added to the `plot`function which offers this option.

- In `test-kb.test.R`, for the current tests, we verify that the output object has the expected class and we verify the correctness of the tests by checking that the tests do not reject or reject the null hypothesis as expected. Hence, the correctness of the tests is verified indirectly. We have specified this in the comment of the G5.4 standard.
