### 2025-01-31 3c699ca
Update test-kb.test.R
-update G5.4 standard
