### 2025-01-31 03cbdf7
Deploying to gh-pages from @ giovsaraceno/QuadratiK-package@3c699caa2590f32892e6b4133aad5e1c7f2a81c0 🚀
