### 2025-02-02 #632 @maurolepore
Dear @giovsaraceno this is to mark the start of my EiC rotation. I'm reviewing all open issues and noting what I see.

I see this submission is a bit ahead of it's labels. The  labels suggest the reviews are pending but they are both done. Also the author responded to both review [here](https://github.com/ropensci/software-review/issues/632#issuecomment-2493060795) and [here](https://github.com/ropensci/software-review/issues/632#issuecomment-2627808858).

Next steps:

- [ ] @kasselhingee, how many hours do you think it took you to review this package?
- [ ] @kasselhingee please see https://github.com/ropensci/software-review/issues/632#issuecomment-2493060795 and indicate if you request further changes, or approve [using this template](https://devguide.ropensci.org/approval2template.html).
- [ ] @emitanaka please see https://github.com/ropensci/software-review/issues/632#issuecomment-2627808858 and indicate if you request further changes, or approve [using this template](https://devguide.ropensci.org/approval2template.html).
- [ ] @emitanaka, I see you're not only a reviewer but also the editor. Once @kasselhingee shares the time estimate, could you please record his review with:

As an example I'll do it below to record your own review.
