### 2025-02-03 #632 @emitanaka
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 10.5 hours
