### 2025-02-03 #632 @ropensci-review-bot
Approved! Thanks @giovsaraceno for submitting and @kasselhingee, @emitanaka for your reviews! :grin:
