### 2025-02-03 #632 @giovsaraceno
Dear @maurolepore,

thanks for accepting the package in ropensci. I have now followed the to-dos listed in the last comment by the ropensci-review-bot. In particular,
- I transfered the repository to rOpenSci, and this has been confirmed.
- I deleted the code of conduct.
- I have changed all the links in the README and DESCRIPTION files pointing to the ropensci repository.
- I fixed the links for the budges.
- I run the codemeta code.
- Added the installation method via ropensci.

I tried to follow the steps for changing the automatic deployment of pkgdown. At the moment the webpage at the new address is not created. I would like to ask if I have just wait or if there is something missing.
