### 2025-02-03 #632 @maurolepore
@giovsaraceno, great! The credit goes to @emitanaka who might be able to help debug the website?

But while I'm here ... I see the URL has an odd `-package` suffix. Maybe something to explore?

Compare: https://github.com/orgs/ropensci/repositories
